### 2025-09-29 #9103 @editorialbot
[PRE REVIEW]: Raptor: A Python Library for Porosity Predictions in Additive Manufacturing
**Submitting author:** @subravvr (<a href="http://orcid.org/0000-0002-9205-8212">Vamsi Subraveti</a>)
**Repository:** https://github.com/ORNL-MDF/Raptor/
**Branch with paper.md** (empty if default branch): joss-2025
**Version:** 1.0.0
