+++ doc/paper.bib
@@ -5,7 +5,7 @@ volume = {20},
 pages = {2109-2172},
 year = {2022},
 issn = {2238-7854},
-doi = {https://doi.org/10.1016/j.jmrt.2022.07.121},
+doi = {10.1016/j.jmrt.2022.07.121},
 url = {https://www.sciencedirect.com/science/article/pii/S2238785422011607},
 author = {Sohini Chowdhury and N. Yadaiah and Chander Prakash and Seeram Ramakrishna and Saurav Dixit and Lovi Raj Gupta and Dharam Buddhi},
 keywords = {Additive manufacturing, Defects, Feedstock materials, Numerical modeling, Properties, Selective laser melting},
@@ -19,7 +19,7 @@ volume = {158},
 pages = {106737},
 year = {2022},
 issn = {0142-1123},
-doi = {https://doi.org/10.1016/j.ijfatigue.2022.106737},
+doi = {10.1016/j.ijfatigue.2022.106737},
 url = {https://www.sciencedirect.com/science/article/pii/S0142112322000184},
 author = {Jaime Berez and Luke Sheridan and Christopher Saldaña},
 keywords = {Laser powder bed fusion, Additive manufacturing, Defect, Probabilistic fatigue, El-Haddad},
@@ -33,7 +33,7 @@ volume = {187},
 pages = {108428},
 year = {2024},
 issn = {0142-1123},
-doi = {https://doi.org/10.1016/j.ijfatigue.2024.108428},
+doi = {10.1016/j.ijfatigue.2024.108428},
 url = {https://www.sciencedirect.com/science/article/pii/S014211232400286X},
 author = {Tharun Reddy and Austin Ngo and Justin P. Miner and Christian Gobert and Jack L. Beuth and Anthony D. Rollett and John J. Lewandowski and Sneha P. Narra},
 keywords = {Additive manufacturing, Laser beam powder bed fusion, Defects, Porosity, Fatigue},
@@ -47,7 +47,7 @@ volume = {95},
 pages = {104506},
 year = {2024},
 issn = {2214-8604},
-doi = {https://doi.org/10.1016/j.addma.2024.104506},
+doi = {10.1016/j.addma.2024.104506},
 url = {https://www.sciencedirect.com/science/article/pii/S2214860424005529},
 author = {Justin P. Miner and Austin Ngo and Christian Gobert and Tharun Reddy and John J. Lewandowski and Anthony D. Rollett and Jack Beuth and Sneha Prabha Narra},
 keywords = {Powder bed fusion-laser beam, X-ray micro CT, Lack-of-fusion, Melt pool variability, Fatigue life},
@@ -61,7 +61,7 @@ volume = {14},
 pages = {39-48},
 year = {2017},
 issn = {2214-8604},
-doi = {https://doi.org/10.1016/j.addma.2016.12.001},
+doi = {10.1016/j.addma.2016.12.001},
 url = {https://www.sciencedirect.com/science/article/pii/S2214860416300471},
 author = {Ming Tang and P. Chris Pistorius and Jack L. Beuth},
 keywords = {Powder bed fusion, Porosity, Simulation, Melt pool, Lack of fusion},
@@ -75,7 +75,7 @@ volume = {36},
 pages = {101552},
 year = {2020},
 issn = {2214-8604},
-doi = {https://doi.org/10.1016/j.addma.2020.101552},
+doi = {10.1016/j.addma.2020.101552},
 url = {https://www.sciencedirect.com/science/article/pii/S2214860420309246},
 author = {Jerard V. Gordon and Sneha P. Narra and Ross W. Cunningham and He Liu and Hangman Chen and Robert M. Suter and Jack L. Beuth and Anthony D. Rollett},
[…]
