### 2025-09-30 16c2535
Correct author ORCID IDs in paper.md
