### 2025-09-30 02db57b
Simplify installation instructions in README
clarified installation.
