### 2025-10-01 #9103 @kyleniemeyer
@subravvr everyone who has an ORCID needs one - if someone doesn't have one, leave it blank rather than giving an invalid ORCID.
