### 2025-10-06 #9103 @kyleniemeyer
@subravvr can you address the remaining ORCID issue? If someone does not have one, just remove that field—don't put an invalid ORCID.
