### 2025-10-07 7075e72
Update ORCID for Çağlar Oskay
Removed ORCID number for Çağlar Oskay.
