### 2025-11-11 bb66b15
testing suite for api endpoints and core.
