+++ tests/test_api.py
@@ -1,0 +1,532 @@
+# =============================================================================
+# Copyright (c) 2025 Oak Ridge National Laboratory
+#
+# All rights reserved.
+#
+# This file is part of Raptor.
+#
+# For details, see the top-level LICENSE file at:
+# https://github.com/ORNL-MDF/Raptor/LICENSE
+# =============================================================================
+"""
+Test suite for raptor.api module.
+
+This module contains unit tests for all public API functions in the raptor.api module,
+including grid creation, path vector generation, spectral component computation,
+melt pool creation, porosity computation, and VTK output generation.
+"""
+
+import pytest
+import numpy as np
+import tempfile
+import os
+from pathlib import Path
+from typing import List, Dict, Any
+
+# Import the module under test
+from raptor.api import (
+    create_grid,
+    create_path_vectors,
+    compute_spectral_components,
+    create_melt_pool,
+    compute_porosity,
+    write_vtk,
+    compute_morphology,
+    write_morphology,
+)
+from raptor.structures import Grid, PathVector, MeltPool
+
+
+# =============================================================================
+# Fixtures
+# =============================================================================
+
+@pytest.fixture
+def sample_bound_box():
+    """Fixture providing a sample bounding box for testing."""
+    return np.array([
+        [0.0, 0.0, 0.0],  # min point
+        [1.0, 1.0, 0.5]   # max point
+    ])
+
+
+@pytest.fixture
+def sample_voxel_resolution():
+    """Fixture providing a sample voxel resolution."""
+    return 0.01
+
+@pytest.fixture
+def sample_path_vectors():
+    """Fixture providing sample path vectors."""
+    start_point = np.array([0.0, 0.0, 0.0])
+    end_point = np.array([1.0, 1.0, 0.0])
+    start_time = 0.0
+    end_time = 1.0
+    path_vector = PathVector(
+        start_point=start_point,
+        end_point=end_point,
+        start_time=start_time,
+        end_time=end_time
+    )
+    return [path_vector]
+
+@pytest.fixture
+def sample_process_parameters():
+    """Fixture providing sample process parameters."""
+    return {
+        'power': 200.0,
+        'scan_speed': 1.0,
+        'hatch_spacing': 0.1,
+        'layer_height': 0.05,
+        'rotation': 67.0,
+        'scan_extension': 0.1,
+        'extra_layers': 0
+    }
+
+
+@pytest.fixture
+def sample_time_series_data():
+    """Fixture providing sample time series data for melt pool."""
+    t = np.linspace(0, 1, 100)
+    values = 0.0001 + 0.00002 * np.sin(2 * np.pi * 5 * t)
+    return np.column_stack([t, values])
+
+
+@pytest.fixture
+def sample_spectral_components():
+    """Fixture providing sample spectral components."""
+    # Format: [amplitude, frequency, phase]
+    return np.array([
+        [0.0001, 0.0, 0.0],      # mode 0 (mean)
+        [0.00002, 5.0, 0.0],     # mode 1
+        [0.00001, 10.0, np.pi/2] # mode 2
+    ], dtype=np.float64)
+
+
+@pytest.fixture
+def sample_melt_pool_dict(sample_time_series_data):
+    """Fixture providing a sample melt pool dictionary."""
+    return {
+        'width': (sample_time_series_data, 3, 1.0, 2.0),
+        'depth': (sample_time_series_data, 3, 1.0, 2.0),
+        'height': (sample_time_series_data, 3, 1.0, 2.0)
+    }
+
+
+@pytest.fixture
+def temp_output_dir():
+    """Fixture providing a temporary directory for output files."""
+    with tempfile.TemporaryDirectory() as tmpdir:
+        yield Path(tmpdir)
+
+
+# =============================================================================
[…]
