### 2025-11-11 2594837
updated regionprops_table call from scikit-image
