### 2025-11-11 d2ee6df
cleaned up warnings re: single mode
