+++ tests/test_api.py
@@ -41,19 +41,18 @@ from raptor.structures import Grid, PathVector, MeltPool
 # Fixtures
 # =============================================================================
 
+
 @pytest.fixture
 def sample_bound_box():
     """Fixture providing a sample bounding box for testing."""
-    return np.array([
-        [0.0, 0.0, 0.0],  # min point
-        [1.0, 1.0, 0.5]   # max point
-    ])
+    return np.array([[0.0, 0.0, 0.0], [1.0, 1.0, 0.5]])  # min point  # max point
 
 
 @pytest.fixture
 def sample_voxel_resolution():
     """Fixture providing a sample voxel resolution."""
     return 0.01
+
 
 @pytest.fixture
 def sample_path_vectors():
@@ -66,21 +65,22 @@ def sample_path_vectors():
         start_point=start_point,
         end_point=end_point,
         start_time=start_time,
-        end_time=end_time
+        end_time=end_time,
     )
     return [path_vector]
+
 
 @pytest.fixture
 def sample_process_parameters():
     """Fixture providing sample process parameters."""
     return {
-        'power': 200.0,
-        'scan_speed': 1.0,
-        'hatch_spacing': 0.1,
-        'layer_height': 0.05,
-        'rotation': 67.0,
-        'scan_extension': 0.1,
-        'extra_layers': 0
+        "power": 200.0,
+        "scan_speed": 1.0,
+        "hatch_spacing": 0.1,
+        "layer_height": 0.05,
+        "rotation": 67.0,
+        "scan_extension": 0.1,
+        "extra_layers": 0,
     }
 
 
@@ -96,20 +96,23 @@ def sample_time_series_data():
 def sample_spectral_components():
     """Fixture providing sample spectral components."""
     # Format: [amplitude, frequency, phase]
-    return np.array([
-        [0.0001, 0.0, 0.0],      # mode 0 (mean)
-        [0.00002, 5.0, 0.0],     # mode 1
-        [0.00001, 10.0, np.pi/2] # mode 2
-    ], dtype=np.float64)
+    return np.array(
+        [
+            [0.0001, 0.0, 0.0],  # mode 0 (mean)
+            [0.00002, 5.0, 0.0],  # mode 1
+            [0.00001, 10.0, np.pi / 2],  # mode 2
+        ],
+        dtype=np.float64,
+    )
 
 
 @pytest.fixture
 def sample_melt_pool_dict(sample_time_series_data):
     """Fixture providing a sample melt pool dictionary."""
     return {
-        'width': (sample_time_series_data, 3, 1.0, 2.0),
-        'depth': (sample_time_series_data, 3, 1.0, 2.0),
-        'height': (sample_time_series_data, 3, 1.0, 2.0)
+        "width": (sample_time_series_data, 3, 1.0, 2.0),
+        "depth": (sample_time_series_data, 3, 1.0, 2.0),
+        "height": (sample_time_series_data, 3, 1.0, 2.0),
     }
 
 
@@ -124,28 +127,32 @@ def temp_output_dir():
 # Tests for create_grid
 # =============================================================================
 
+
 class TestCreateGrid:
     """Test cases for the create_grid function."""
-    
-    def test_create_grid_with_bound_box(self, sample_voxel_resolution, sample_bound_box):
+
+    def test_create_grid_with_bound_box(
+        self, sample_voxel_resolution, sample_bound_box
+    ):
         """Test grid creation with a bounding box."""
         grid = create_grid(sample_voxel_resolution, bound_box=sample_bound_box)
-        
+
         assert isinstance(grid, Grid)
         assert grid.resolution == sample_voxel_resolution
         assert grid.origin.shape == (3,)
[…]
