### 2025-11-13 769c8f1
Testsuite (#21)
* testing suite for api endpoints and core.

* updated regionprops_table call from scikit-image

* cleaned up warnings re: single mode

* add pore to supress warning in pytest. fit formatting

---------

Co-authored-by: Vamsi Subraveti <vis@ornl.gov>
Co-authored-by: John <colemanjs@ornl.gov>
