+++ tests/test_core.py
@@ -1,0 +1,625 @@
+# =============================================================================
+# Copyright (c) 2025 Oak Ridge National Laboratory
+#
+# All rights reserved.
+#
+# This file is part of Raptor.
+#
+# For details, see the top-level LICENSE file at:
+# https://github.com/ORNL-MDF/Raptor/LICENSE
+# =============================================================================
+"""
+Test suite for raptor.core module.
+
+This module contains unit tests for the core computation functions including
+geometric containment testing, melt mask computation, and the main parallelized
+implicit melt mask calculation.
+"""
+
+import pytest
+import numpy as np
+from typing import List
+from unittest.mock import Mock, MagicMock
+
+# Import the module under test
+from raptor.core import is_inside, compute_melt_mask, compute_melt_mask_implicit
+from raptor.structures import MeltPool, PathVector
+
+
+# =============================================================================
+# Fixtures
+# =============================================================================
+
+
+@pytest.fixture
+def sample_melt_pool():
+    """Fixture providing a sample MeltPool object."""
+    width_osc = np.array(
+        [[0.0001, 0.0, 0.0], [0.00002, 5.0, 0.0], [0.00001, 10.0, 0.0]],
+        dtype=np.float64,
+    )
+
+    depth_osc = np.array(
+        [[0.00008, 0.0, 0.0], [0.00001, 5.0, 0.0], [0.00001, 10.0, 0.0]],
+        dtype=np.float64,
+    )
+
+    height_osc = np.array(
+        [[0.00006, 0.0, 0.0], [0.00001, 5.0, 0.0], [0.00001, 10.0, 0.0]],
+        dtype=np.float64,
+    )
+
+    return MeltPool(
+        width_oscillations=width_osc,
+        depth_oscillations=depth_osc,
+        height_oscillations=height_osc,
+        width_max=0.00012,
+        depth_max=0.0001,
+        height_max=0.00008,
+        width_shape_factor=2.0,
+        height_shape_factor=2.0,
+        depth_shape_factor=2.0,
+        enable_random_phases=False,
+    )
+
+
+@pytest.fixture
+def sample_path_vector():
+    """Fixture providing a sample PathVector object."""
+    start = np.array([0.0, 0.0, 0.0], dtype=np.float64)
+    end = np.array([0.001, 0.0, 0.0], dtype=np.float64)
+
+    path_vec = PathVector(
+        start_point=start, end_point=end, start_time=0.0, end_time=0.001
+    )
+    path_vec.set_coordinate_frame()
+
+    return path_vec
+
+
+@pytest.fixture
+def sample_path_vectors(sample_path_vector):
+    """Fixture providing a list of sample PathVector objects."""
+    vectors = []
+
+    # Create multiple parallel vectors
+    for i in range(3):
+        start = np.array([0.0, i * 0.0001, 0.0], dtype=np.float64)
+        end = np.array([0.001, i * 0.0001, 0.0], dtype=np.float64)
+
+        vec = PathVector(
+            start_point=start,
+            end_point=end,
+            start_time=i * 0.001,
+            end_time=(i + 1) * 0.001,
+        )
+        vec.set_coordinate_frame()
+        vectors.append(vec)
+
+    return vectors
+
+
+@pytest.fixture
+def sample_voxels():
+    """Fixture providing a sample voxel array."""
+    # Create a small grid of voxels
+    x = np.linspace(0, 0.001, 10)
+    y = np.linspace(0, 0.0005, 5)
+    z = np.linspace(0, 0.0002, 3)
+
+    xx, yy, zz = np.meshgrid(x, y, z, indexing="ij")
+
[…]
