### 2025-11-25 #9103 @jgostick
Hey @subravvr, the paper WAY TOO LONG.  It has over 3000 words not including references. JOSS is pretty strict about keeping this (well) below 1000 words.  The reason is that we want reviewers to look at your code very closely, and not also have to wade through a full length research article with tons of theory.  In other words, JOSS does not want to be responsible for making sure the paper is theoretically correct and rigorous, only the software.

So, you need to do a complete rewrite of the paper to basically outline the package, maybe mention a few of the modules and capabilities in rough outline.  You can put the theory and examples on your website/docs.

Sorry, but I have just come off of 2 reviewers where the papers were too long and it was problematic, so I insist to avoid this here.
