### 2025-12-10 e038c2c
trimmed paper.
