+++ doc/paper.md
@@ -31,59 +31,14 @@ header-includes:
 
 Raptor is an efficient Python library for simulating stochastic lack-of-fusion (sLoF) defects in additive manufacturing (AM) processes. These defects arise from stochastic variations in the melt pool boundary leading to undermelting from insufficient overlap between adjacent melt pools or successive layers. Performance variability of AM parts is a pressing challenge in qualification and certification of AM parts; this is in part due to the poorly understood formation rate and statistics of sLoF defects. Raptor is designed to capture the explicit morphologies of sLoF defects and their statistics to accelerate qualification and certification efforts of AM parts in critical applications.
 
-Laser powder bed fusion (LPBF) is a metal AM technique which can produce complex, bespoke parts that are challenging to fabricate using traditional methods. Such components are commonly seen in critical applications, such as topology-optimized parts in the aerospace and defense industries [@chowdhury_review_2019]. A prevalent challenge in the adoption of these parts is their qualification and certification (Q+C) regarding mechanical performance. LPBF has been shown to produce parts with 
-poor mechanical performance due to inherent process stochasticity effects on the spread of fatigue life [@berez_fatiguevariation_2022; @reddy_fatigue_2024]. This probability of poor performance remains as a barrier in the Q+C efforts of LPBF-built parts. Recent studies have revealed the role of stochastic defects, specifically sLoF, in reducing performance characteristics of LPBF parts [@miner_lof_2024]. To understand the formation and morphology of sLoF in an already optimized printing regime, a model-based approach must be used to query this space; an experimental approach would result in prohibitively high material costs. For this tangible impact on Q+C activites, Raptor offers a new state-of-the-art in rapid explicit sLoF defect structure prediction. 
-
 
 # Statement of need
-Lack-of-fusion (LoF) defects are a type of defect in LPBF that result from local undermelting caused by insufficient overlap between adjacent melt pools or successive layers [@tang_lof_2017; @gordon_defectstructure_2020]. The analytical prediction of LoF defects commonly relate the melt pool dimensions to spacing parameters chosen before the build [@tang_lof_2017]. Assuming the melt pool is elliptical and defined by its width $W$ and depth $D$, and adjacent melt pools are separated by hatch spacing $H$ and layer thickness $L$, a common criterion used to determine the formation of LoF is given by
 
-\begin{equation}
-\left( \frac{L}{D} \right)^2 + \left( \frac{H}{W} \right)^2 \leq 1
-\end{equation}
-
[…]
