+++ doc/paper.md
@@ -34,7 +34,11 @@ Raptor is an efficient Python library for simulating stochastic lack-of-fusion (
 
 # Statement of need
 
-In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. The melt pool dimensions are treated as a stochastic process informed by user-supplied data. The stochastic melt pool dimensions are efficiently sampled via a truncated cosine expansion of the full Fourier basis. The locally computed dimensions from the cosine expansion are used to define a Lame curve, which is then used to mask a voxelized grid. The implementation leverages a point-parallel approach, where the independence of each voxel in the domain with any other voxel allows for parallel processing of the masking operation. This paradigm makes probabilistic analyses of explicit sLoF defect structures tractable. Raptor is able to process representative volume elements (RVEs) of edge size 0.5 mm with a resolution of 2.5 µm (8,000,000 voxels) within 1-4 seconds on a local workstation; this RVE simulation can be repeatedly queried to construct an ensemble of 1000 sLoF realizations in under an hour, or an equivalent scanned volume of 125 mm$^3$. The scale of this RVE ensemble is able to capture rare sLoF defect events and is targeted toward simulating defects responsible for deleterious part performance within the optimized printing regime [@reddy_fatigue_2024].
+In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. 
+
+The current modeling landscape for sLoF prediction is fairly sparse; initial work from Tang et al. 2017.
+
+The melt pool dimensions are treated as a stochastic process informed by user-supplied data. The stochastic melt pool dimensions are efficiently sampled via a truncated cosine expansion of the full Fourier basis. The locally computed dimensions from the cosine expansion are used to define a Lame curve, which is then used to mask a voxelized grid. The implementation leverages a point-parallel approach, where the independence of each voxel in the domain with any other voxel allows for parallel processing of the masking operation. This paradigm makes probabilistic analyses of explicit sLoF defect structures tractable. Raptor is able to process representative volume elements (RVEs) of edge size 0.5 mm with a resolution of 2.5 µm (8,000,000 voxels) within 1-4 seconds on a local workstation; this RVE simulation can be repeatedly queried to construct an ensemble of 1000 sLoF realizations in under an hour, or an equivalent scanned volume of 125 mm$^3$. The scale of this RVE ensemble is able to capture rare sLoF defect events and is targeted toward simulating defects responsible for deleterious part performance within the optimized printing regime [@reddy_fatigue_2024].
 
 # Software features
 
