+++ doc/paper.bib
@@ -211,6 +211,21 @@ doi = {10.1007/s40192-024-00347-5},
 author = {Vamsi Subraveti and Brodan Richter and Sai Yeratapally and Caglar Oskay}
 }
 
+@article{subraveti_process_2025,
+	title = {Process {Uncertainty} {Analysis} of {Stochastic} {Lack}-of-{Fusion} {Defects} in {Laser} {Powder} {Bed} {Fused} {Inconel} 718},
+	volume = {14},
+	issn = {2193-9772},
+	url = {https://doi.org/10.1007/s40192-025-00421-6},
+	doi = {10.1007/s40192-025-00421-6},
+	abstract = {This manuscript introduces a computational framework to simulate the generation of stochastic lack-of-fusion (LoF) defects produced by random fluctuations in the melt pool geometry during laser powder bed fusion (LPBF). The framework facilitates the analysis of the relationship between the manufacturing process parameters and the resulting stochasticity in lack-of-fusion defects. Several sources of uncertainty in the LPBF process contribute to this stochasticity, and this analysis quantifies the aforementioned relationship. A spectral matching algorithm was developed to statistically reproduce experimentally observed melt pool fluctuations obtained from image segmentation of scan tracks. These fluctuations were coupled with an analytical mean melt pool geometry model to estimate dynamic melt pool behavior in LPBF IN718. The dynamic melt pool histories were employed to simulate LoF defects. The volume fraction of simulated LoF defects was validated against volume fractions from IN718 experimental builds reported in the literature. Parameter combinations in the power–velocity (P–V) and layer thickness–hatch (L–H) spaces were sampled with dynamic melt pool histories; statistical volume element (SVE) simulations were then executed, and the distributions of defect volume fraction, equivalent diameter and sphericity were constructed. The extreme values of both equivalent diameter and sphericity were noted to be present even near fully dense conditions in the P–V space. Stochastic LoF defects were found to occur within the accepted criterion curve in the L–H space, indicating the impact of melt pool fluctuations on the generation of stochastic LoF in nominally defect-free regimes. A sub-optimal process regime was identified in the L–H space. The proposed framework offers a way to assess the impact of parameter selection on LoF defect distributions to further advance the design, qualification and certification of LPBF parts.},
+	number = {4},
+	journal = {Integrating Materials and Manufacturing Innovation},
+	author = {Subraveti, Vamsi and Richter, Brodan and Pribe, Joshua D. and Weber, George R. and Oskay, Caglar},
+	month = dec,
+	year = {2025},
+	pages = {657--678},
+}
+
 
 
 
