+++ doc/paper.md
@@ -36,7 +36,7 @@ Raptor is an efficient Python library for simulating stochastic lack-of-fusion (
 
 In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. 
 
-The current modeling landscape for sLoF prediction is fairly sparse; initial work from Tang et al. 2017.
+The current modeling landscape for sLoF prediction is fairly sparse; initial work from Tang et al. 2017 provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction by Subraveti et al. 2024, still utilizing the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, Richter et al. 2025 introduced uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model [@richter_analyticallof_2025]. The current state of the art sLoF model in Subraveti et al. 2025 utilizes a Fourier-based expansion of the melt pool dimensions to account for temporal variability and solves for explicit overlap-driven sLoF geometries [@subraveti_sma_2025, @subraveti_process_2025].
 
 The melt pool dimensions are treated as a stochastic process informed by user-supplied data. The stochastic melt pool dimensions are efficiently sampled via a truncated cosine expansion of the full Fourier basis. The locally computed dimensions from the cosine expansion are used to define a Lame curve, which is then used to mask a voxelized grid. The implementation leverages a point-parallel approach, where the independence of each voxel in the domain with any other voxel allows for parallel processing of the masking operation. This paradigm makes probabilistic analyses of explicit sLoF defect structures tractable. Raptor is able to process representative volume elements (RVEs) of edge size 0.5 mm with a resolution of 2.5 µm (8,000,000 voxels) within 1-4 seconds on a local workstation; this RVE simulation can be repeatedly queried to construct an ensemble of 1000 sLoF realizations in under an hour, or an equivalent scanned volume of 125 mm$^3$. The scale of this RVE ensemble is able to capture rare sLoF defect events and is targeted toward simulating defects responsible for deleterious part performance within the optimized printing regime [@reddy_fatigue_2024].
 
