+++ doc/paper.md
@@ -29,22 +29,31 @@ header-includes:
 
 # Summary
 
-Raptor is an efficient Python library for simulating stochastic lack-of-fusion (sLoF) defects in additive manufacturing (AM) processes. These defects arise from stochastic variations in the melt pool boundary leading to undermelting from insufficient overlap between adjacent melt pools or successive layers. Performance variability of AM parts is a pressing challenge in qualification and certification of AM parts; this is in part due to the poorly understood formation rate and statistics of sLoF defects. Raptor is designed to capture the explicit morphologies of sLoF defects and their statistics to accelerate qualification and certification efforts of AM parts in critical applications.
-
+Raptor is an efficient Python library for simulating stochastic lack-of-fusion (sLoF) defects in additive manufacturing (AM) processes. These defects arise from stochastic variations in the melt pool boundary leading to undermelting from insufficient overlap between adjacent melt pools or successive layers [@khairallah_physics_2016, @grasso_insitu_2017]. Performance variability of AM parts is a pressing challenge in qualification and certification of AM parts; this is in part due to the poorly understood formation rate and statistics of sLoF defects. Raptor is designed to capture the explicit morphologies of sLoF defects and their statistics to accelerate qualification and certification efforts of AM parts in critical applications.
 
 # Statement of need
 
-In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. 
+Metal AM processes show great promise in advancing manufacturing capabilities across a variety of industries. However, the quantification of uncertainty in the desired properties of AM parts is an ongoing challenge which spans many disciplines of engineering. A key driver in this challenge is the modeling of explicit sLoF defect geometries and their occurrence rates [@reddy_fatigue_2024, @berez_fatiguevariation_2022]. Experimental observations of sLoF show that these defects persist well into the previously determined optimal processing regime; their occurrence, however, is sparse [@miner_lof_2024]. The sparsity of the sLoF defects coupled with their impact on desirable properties necessitates a focused modeling effort on simulation, prediction, and mitigation of these defects to accelerate metal AM adoption throughout industry.   
+
+# State of the field
 
 The current modeling landscape for sLoF prediction is fairly sparse; initial work from Tang et al. 2017 provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction by Subraveti et al. 2024, still utilizing the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, Richter et al. 2025 introduced uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model [@richter_analyticallof_2025]. The current state of the art sLoF model in Subraveti et al. 2025 utilizes a Fourier-based expansion of the melt pool dimensions to account for temporal variability and solves for explicit overlap-driven sLoF geometries [@subraveti_sma_2025, @subraveti_process_2025].
 
[…]
