+++ doc/paper.md
@@ -37,7 +37,7 @@ Metal AM processes show great promise in advancing manufacturing capabilities ac
 
 # State of the field
 
-The current modeling landscape for sLoF prediction is fairly sparse; initial work from Tang et al. 2017 provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction by Subraveti et al. 2024, still utilizing the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, Richter et al. 2025 introduced uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model [@richter_analyticallof_2025]. The current state of the art sLoF model in Subraveti et al. 2025 utilizes a Fourier-based expansion of the melt pool dimensions to account for temporal variability and solves for explicit overlap-driven sLoF geometries [@subraveti_sma_2025, @subraveti_process_2025].
+The current modeling landscape for sLoF prediction is fairly sparse; initial work provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction, continuing to utilize the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model has been introduced [@richter_analyticallof_2025]. The current state of the art sLoF model builds on this approach with a Fourier-based expansion of the melt pool dimensions to account for temporal variability [@subraveti_sma_2025, @subraveti_process_2025]. This model solves for explicit melt pool overlaps, yielding sLoF geometries resultant from melt pool fluctuations.
 
 # Software design
 
