+++ doc/paper.md
@@ -29,11 +29,11 @@ header-includes:
 
 # Summary
 
-Raptor is an efficient Python library for simulating stochastic lack-of-fusion (sLoF) defects in additive manufacturing (AM) processes. These defects arise from stochastic variations in the melt pool boundary leading to undermelting from insufficient overlap between adjacent melt pools or successive layers [@khairallah_physics_2016 @grasso_insitu_2017]. Performance variability of AM parts is a pressing challenge in qualification and certification of AM parts; this is in part due to the poorly understood formation rate and statistics of sLoF defects. Raptor is designed to capture the explicit morphologies of sLoF defects and their statistics to accelerate qualification and certification efforts of AM parts in critical applications.
+Raptor is an efficient Python library for simulating stochastic lack-of-fusion (sLoF) defects in additive manufacturing (AM) processes. These defects arise from stochastic variations in the melt pool boundary leading to undermelting from insufficient overlap between adjacent melt pools or successive layers [@khairallah_physics_2016;@grasso_insitu_2017]. Performance variability of AM parts is a pressing challenge in qualification and certification of AM parts; this is in part due to the poorly understood formation rate and statistics of sLoF defects. Raptor is designed to capture the explicit morphologies of sLoF defects and their statistics to accelerate qualification and certification efforts of AM parts in critical applications.
 
 # Statement of need
 
-Metal AM processes show great promise in advancing manufacturing capabilities across a variety of industries. However, the quantification of uncertainty in the desired properties of AM parts is an ongoing challenge which spans many disciplines of engineering. A key driver in this challenge is the modeling of explicit sLoF defect geometries and their occurrence rates [@reddy_fatigue_2024, @berez_fatiguevariation_2022]. Experimental observations of sLoF show that these defects persist well into the previously determined optimal processing regime; their occurrence, however, is sparse [@miner_lof_2024]. The sparsity of the sLoF defects coupled with their impact on desirable properties necessitates a focused modeling effort on simulation, prediction, and mitigation of these defects to accelerate metal AM adoption throughout industry.   
+Metal AM processes show great promise in advancing manufacturing capabilities across a variety of industries. However, the quantification of uncertainty in the desired properties of AM parts is an ongoing challenge which spans many disciplines of engineering. A key driver in this challenge is the modeling of explicit sLoF defect geometries and their occurrence rates [@reddy_fatigue_2024;@berez_fatiguevariation_2022]. Experimental observations of sLoF show that these defects persist well into the previously determined optimal processing regime; their occurrence, however, is sparse [@miner_lof_2024]. The sparsity of the sLoF defects coupled with their impact on desirable properties necessitates a focused modeling effort on simulation, prediction, and mitigation of these defects to accelerate metal AM adoption throughout industry.   
 
 # State of the field
 
@@ -43,11 +43,11 @@ The current modeling landscape for sLoF prediction is fairly sparse; initial wor
 
 In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. The novel parallelized algorithm we present results multiple order-of-magnitude efficiency gain over the previous state-of-the-art serially timestepped approach. In Raptor, the melt pool dimensions are treated as a stochastic process informed by user-supplied data. These stochastic processes are efficiently sampled via a truncated cosine expansion of the full Fourier basis computed from the user-supplied data. The locally computed dimensions from the cosine expansion are used to define a Lamé curve, which is then used to mask a voxelized grid representing the printed volume of interest.
 
[…]
