### 2026-01-05 431e3af
fixing some small sentences
