+++ doc/paper.md
@@ -33,11 +33,11 @@ Raptor is an efficient Python library for simulating stochastic lack-of-fusion (
 
 # Statement of need
 
-Metal AM processes show great promise in advancing manufacturing capabilities across a variety of industries. However, the quantification of uncertainty in the desired properties of AM parts is an ongoing challenge which spans many disciplines of engineering. A key driver in this challenge is the modeling of explicit sLoF defect geometries and their occurrence rates [@reddy_fatigue_2024;@berez_fatiguevariation_2022]. Experimental observations of sLoF show that these defects persist well into the previously determined optimal processing regime; their occurrence, however, is sparse [@miner_lof_2024]. The sparsity of the sLoF defects coupled with their impact on desirable properties necessitates a focused modeling effort on simulation, prediction, and mitigation of these defects to accelerate metal AM adoption throughout industry.   
+Metal AM processes show great promise in advancing manufacturing capabilities across a variety of industries. However, the quantification of uncertainty in the desired properties of AM parts is an ongoing challenge which spans many disciplines of engineering. A key driver in this challenge is the modeling of explicit sLoF defect geometries and their occurrence rates [@reddy_fatigue_2024;@berez_fatiguevariation_2022]. Experimental observations of sLoF show that these defects persist well into the previously determined optimal processing regime; their occurrence, however, is sparse [@miner_lof_2024]. The sparsity of the sLoF defects coupled with their impact on desirable properties necessitates a focused, collaborative modeling effort for the simulation, prediction, and mitigation of these defects to accelerate metal AM adoption throughout industry. 
 
 # State of the field
 
-The current modeling landscape for sLoF prediction is fairly sparse; initial work provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction, continuing to utilize the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model has been introduced [@richter_analyticallof_2025]. The current state of the art sLoF model builds on this approach with a Fourier-based expansion of the melt pool dimensions to account for temporal variability [@subraveti_sma_2025;@subraveti_process_2025]. This model solves for explicit melt pool overlaps, yielding sLoF geometries resultant from melt pool fluctuations. One of the key points we aim to address with Raptor is availability; none of the mentioned models are currently publicly available.
+The current modeling landscape for sLoF prediction is fairly sparse; initial work provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction, continuing to utilize the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model has been introduced [@richter_analyticallof_2025]. The current state of the art sLoF model builds on this approach with a Fourier-based expansion of the melt pool dimensions to account for temporal variability [@subraveti_sma_2025;@subraveti_process_2025]. This model solves for explicit melt pool overlaps, yielding sLoF geometries resultant from melt pool fluctuations. 
 
 # Software design
 
