### 2026-01-05 #9103 @subravvr
Hi @jgostick!
I have gone through the requirements and restructured the paper to fit them. I have added a section discussing the current modeling landscape (complete with citations). I have also expounded a bit on the statement of need and added the GenAI section.
Thank you for your patience!
