### 2026-01-09 #9103 @jgostick
The article looks great now, thanks!  I will proceed with finding reviewers.
