### 2026-01-22 #9874 @jgostick
Hi @openjournals/joss-eics, I believe this submission meets the criteria for pre-2026 guidelines.  I have another submission where someone or something magically added the pre-2026 label and also updated the reviewers checklist. I hoping that can happen here too?
