### 2026-02-24 1b479a6
active learning branch, modified rve.py to generate enough rves for a query volume.
