### 2026-03-01 #9874 @jgostick
Hi @renelam.  You can make general comments here, or you can open an issue on their repo for specific things like problems installing or other bugs.  Even suggestions about improving the docs can be done with an issue.
