### 2026-03-02 7704b93
updated example for various hatch spacings
