### 2026-03-03 a59b64b
fixup of structure for active learning
