+++ examples/api_example/active_learning_example.py
@@ -9,83 +9,76 @@ from raptor.api import (
 )
 from raptor.utilities import ScanPathBuilder
 
-# --- GLOBAL SETTINGS ---
-query_volume_mm3 = 1.0
-voxel_resolution_m = 5.0e-6
 
-# --- RVE (Representative Volume Element) Dimensions ---
-rve_min_point = np.array([0.0, 0.0, 0.0])
-rve_max_point = np.array([5.0e-4, 5.0e-4, 5.0e-4])
-rve_bounding_box = np.array([rve_min_point, rve_max_point])
+def compute_mean_and_std(defects: np.ndarray):
+    if len(defects) == 0:
+        y = 0.0
+        yerr = np.nan
+    elif len(defects) == 1:
+        y = defects[0]
+        yerr = defects[0]
+    else:
+        y = np.mean(defects)
+        yerr = np.std(defects, ddof=1) / np.sqrt(len(defects))
 
-# --- Laser & Scan Path Machine Parameters ---
-laser_power_watts = 370.0  # Laser power in Watts
-laser_velocity_m_s = 1.7  # Laser scan speed in meters per second
-powder_layer_thickness_m = 50.0e-6  # 50 microns layer height
-layer_rotation_angle_deg = 67.0  # Rotation of the scan between layers
-scan_extension_distance_m = max(rve_max_point - rve_min_point)
-num_extra_layers = 0
-
-# --- Melt Pool Morphology Parameters ---
-SCRIPT_DIR = Path(__file__).resolve().parent
-melt_pool_data_path = (
-    SCRIPT_DIR / ".." / "data" / "meltPoolData" / "ULI_v1700_theta0_widths.txt"
-)
-base_width_data = read_data(melt_pool_data_path)
-melt_pool_modes = 50
-
-# Shape identifiers defined by Raptor: 1 = Parabola, 2 = Ellipse
-melt_pool_width_scale = 1.0
-melt_pool_depth_scale = 0.8
-melt_pool_height_scale = 0.4
-
-width_shape_type = 2
-depth_shape_type = 1
-height_shape_type = 1
-
-# Create the 3D voxel grid for the RVE
-grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
-
-# Configure the dictionary that defines the 3D melt pool volume
-melt_pool_dict = {
-    "width": (
-        base_width_data,
-        melt_pool_modes,
-        melt_pool_width_scale,
-        width_shape_type,
-    ),
-    "depth": (
-        base_width_data,
-        melt_pool_modes,
-        melt_pool_depth_scale,
-        depth_shape_type,
-    ),
-    "height": (
-        base_width_data,
-        melt_pool_modes,
-        melt_pool_height_scale,
-        height_shape_type,
-    ),
-}
-
-# Generate the moving melt pool with realistic random spatial fluctuations
-melt_pool = create_melt_pool(melt_pool_dict, enable_random_phases=True)
-
-rve_volume_mm3 = (1e3 * (rve_max_point - rve_min_point)).prod()
-print(rve_volume_mm3)
-num_rves = int(np.ceil(query_volume_mm3 / rve_volume_mm3))
-print(f"  -> Simulating {num_rves} RVEs to fill {query_volume_mm3} mm^3...")
+    return y, yerr
 
 
-# --- Function to evaluate hatch spacing ----
-def evaluate_hatch_spacing(hatch_spacing_m: float, requires_jit_warmup: bool) -> float:
-    # Build the laser paths using our machine parameters
+def defects_model(
+    hatch_spacing_m: float, layer_thickness_m: float, query_volume_mm3: float = 1.0
+) -> np.ndarray:
+
+    # Create the 3D voxel grid for the RVE
[…]
