### 2026-03-06 474c6f8
fixup to generalize outputs of interest
