### 2026-03-06 6d43928
fixup to generalize outputs of interest
