### 2026-03-16 35b317b
added morphology writing to api. renamed scan path file in cli
