+++ examples/api_example/rve.py
@@ -17,6 +17,8 @@ from raptor.api import (
     create_grid,
     compute_porosity,
     write_vtk,
+    compute_morphology,
+    write_morphology
 )
 from raptor.utilities import ScanPathBuilder
 
@@ -31,8 +33,8 @@ grid = create_grid(voxel_resolution, bound_box=bound_box)
 # 2. Create path vectors through the representative volume element (RVE)
 power = 370
 velocity = 1.7
-hatch_spacing = 130e-6
-layer_height = 50e-6
+hatch_spacing = 140e-6
+layer_height = 30e-6
 rotation = 67
 scan_extension = max(max_point - min_point)
 extra_layers = 0
@@ -80,5 +82,10 @@ melt_pool = create_melt_pool(melt_pool_dict, enable_random_phases=True)
 # 4. Compute porosity using conic section / superellipse curves for melt pool mask
 porosity = compute_porosity(grid, path_vectors, melt_pool, jit_warmup=True)
 
-# 5. Write porosity field to .VTI
+# 5. Compute morphology
+morphology = compute_morphology(porosity, voxel_resolution, ['area', 'equivalent_diameter_area'])
+write_morphology(morphology, "rve_morphology.csv")
+
+
+# 6. Write porosity field to .VTI
 write_vtk(grid.origin, grid.resolution, porosity, "rve.vti")
