### 2026-03-16 e21eb57
added complex cli example.
