### 2026-03-16 bdb4e81
updated some examples and readme.
