### 2026-03-16 cd9536d
new rve example and updated readme.
