+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_17.txt
@@ -1,0 +1,17 @@
+Mode X(m) Y(m) Z(m) Power(W) tParam
+1.000000 0.000421 -0.000436 0.000510 0.000000 0.000000
+0.000000 0.000936 0.000421 0.000510 370.000000 1.700000
+1.000000 0.000301 -0.000364 0.000510 0.000000 0.000000
+0.000000 0.000816 0.000493 0.000510 370.000000 1.700000
+1.000000 0.000181 -0.000292 0.000510 0.000000 0.000000
+0.000000 0.000696 0.000565 0.000510 370.000000 1.700000
+1.000000 0.000061 -0.000220 0.000510 0.000000 0.000000
+0.000000 0.000576 0.000637 0.000510 370.000000 1.700000
+1.000000 -0.000059 -0.000148 0.000510 0.000000 0.000000
+0.000000 0.000456 0.000709 0.000510 370.000000 1.700000
+1.000000 -0.000179 -0.000076 0.000510 0.000000 0.000000
+0.000000 0.000336 0.000782 0.000510 370.000000 1.700000
+1.000000 -0.000299 -0.000003 0.000510 0.000000 0.000000
+0.000000 0.000216 0.000854 0.000510 370.000000 1.700000
+1.000000 -0.000419 0.000069 0.000510 0.000000 0.000000
+0.000000 0.000096 0.000926 0.000510 370.000000 1.700000
