+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_15.txt
@@ -1,0 +1,17 @@
+Mode X(m) Y(m) Z(m) Power(W) tParam
+1.000000 -0.000362 0.000604 0.000450 0.000000 0.000000
+0.000000 -0.000104 -0.000362 0.000450 370.000000 1.700000
+1.000000 -0.000227 0.000640 0.000450 0.000000 0.000000
+0.000000 0.000032 -0.000326 0.000450 370.000000 1.700000
+1.000000 -0.000092 0.000676 0.000450 0.000000 0.000000
+0.000000 0.000167 -0.000290 0.000450 370.000000 1.700000
+1.000000 0.000043 0.000712 0.000450 0.000000 0.000000
+0.000000 0.000302 -0.000254 0.000450 370.000000 1.700000
+1.000000 0.000179 0.000748 0.000450 0.000000 0.000000
+0.000000 0.000437 -0.000217 0.000450 370.000000 1.700000
+1.000000 0.000314 0.000785 0.000450 0.000000 0.000000
+0.000000 0.000573 -0.000181 0.000450 370.000000 1.700000
+1.000000 0.000449 0.000821 0.000450 0.000000 0.000000
+0.000000 0.000708 -0.000145 0.000450 370.000000 1.700000
+1.000000 0.000584 0.000857 0.000450 0.000000 0.000000
+0.000000 0.000843 -0.000109 0.000450 370.000000 1.700000
