+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_14.txt
@@ -1,0 +1,17 @@
+Mode X(m) Y(m) Z(m) Power(W) tParam
+1.000000 0.000336 0.000952 0.000420 0.000000 0.000000
+0.000000 -0.000452 0.000336 0.000420 370.000000 1.700000
+1.000000 0.000422 0.000842 0.000420 0.000000 0.000000
+0.000000 -0.000366 0.000226 0.000420 370.000000 1.700000
+1.000000 0.000509 0.000731 0.000420 0.000000 0.000000
+0.000000 -0.000279 0.000116 0.000420 370.000000 1.700000
+1.000000 0.000595 0.000621 0.000420 0.000000 0.000000
+0.000000 -0.000193 0.000005 0.000420 370.000000 1.700000
+1.000000 0.000681 0.000511 0.000420 0.000000 0.000000
+0.000000 -0.000107 -0.000105 0.000420 370.000000 1.700000
+1.000000 0.000767 0.000400 0.000420 0.000000 0.000000
+0.000000 -0.000021 -0.000215 0.000420 370.000000 1.700000
+1.000000 0.000853 0.000290 0.000420 0.000000 0.000000
+0.000000 0.000065 -0.000326 0.000420 370.000000 1.700000
+1.000000 0.000940 0.000180 0.000420 0.000000 0.000000
+0.000000 0.000152 -0.000436 0.000420 370.000000 1.700000
