+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_0.txt
@@ -1,0 +1,17 @@
+Mode X(m) Y(m) Z(m) Power(W) tParam
+1.000000 -0.000250 -0.000250 0.000000 0.000000 0.000000
+0.000000 0.000750 -0.000250 0.000000 370.000000 1.700000
+1.000000 -0.000250 -0.000110 0.000000 0.000000 0.000000
+0.000000 0.000750 -0.000110 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000030 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000030 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000170 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000170 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000310 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000310 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000450 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000450 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000590 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000590 0.000000 370.000000 1.700000
+1.000000 -0.000250 0.000730 0.000000 0.000000 0.000000
+0.000000 0.000750 0.000730 0.000000 370.000000 1.700000
