+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_4.txt
@@ -1,17 +1,0 @@
-Mode X(m) Y(m) Z(m) Power(W) tParam
-1.000000 -0.000232 0.000767 0.000160 0.000000 0.000000
-0.000000 -0.000267 -0.000232 0.000160 370.000000 1.700000
-1.000000 -0.000092 0.000762 0.000160 0.000000 0.000000
-0.000000 -0.000127 -0.000237 0.000160 370.000000 1.700000
-1.000000 0.000048 0.000757 0.000160 0.000000 0.000000
-0.000000 0.000013 -0.000242 0.000160 370.000000 1.700000
-1.000000 0.000187 0.000752 0.000160 0.000000 0.000000
-0.000000 0.000153 -0.000247 0.000160 370.000000 1.700000
-1.000000 0.000327 0.000748 0.000160 0.000000 0.000000
-0.000000 0.000293 -0.000252 0.000160 370.000000 1.700000
-1.000000 0.000467 0.000743 0.000160 0.000000 0.000000
-0.000000 0.000432 -0.000257 0.000160 370.000000 1.700000
-1.000000 0.000607 0.000738 0.000160 0.000000 0.000000
-0.000000 0.000572 -0.000262 0.000160 370.000000 1.700000
-1.000000 0.000747 0.000733 0.000160 0.000000 0.000000
-0.000000 0.000712 -0.000266 0.000160 370.000000 1.700000
