+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_5.txt
@@ -1,17 +1,0 @@
-Mode X(m) Y(m) Z(m) Power(W) tParam
-1.000000 -0.000414 0.000008 0.000200 0.000000 0.000000
-0.000000 0.000492 -0.000414 0.000200 370.000000 1.700000
-1.000000 -0.000355 0.000135 0.000200 0.000000 0.000000
-0.000000 0.000551 -0.000288 0.000200 370.000000 1.700000
-1.000000 -0.000296 0.000262 0.000200 0.000000 0.000000
-0.000000 0.000610 -0.000161 0.000200 370.000000 1.700000
-1.000000 -0.000237 0.000389 0.000200 0.000000 0.000000
-0.000000 0.000669 -0.000034 0.000200 370.000000 1.700000
-1.000000 -0.000178 0.000516 0.000200 0.000000 0.000000
-0.000000 0.000729 0.000093 0.000200 370.000000 1.700000
-1.000000 -0.000119 0.000643 0.000200 0.000000 0.000000
-0.000000 0.000788 0.000220 0.000200 370.000000 1.700000
-1.000000 -0.000059 0.000769 0.000200 0.000000 0.000000
-0.000000 0.000847 0.000347 0.000200 370.000000 1.700000
-1.000000 -0.000000 0.000896 0.000200 0.000000 0.000000
-0.000000 0.000906 0.000474 0.000200 370.000000 1.700000
