+++ README.md
@@ -30,7 +30,7 @@ This project is licensed under the BSD 3-Clause [License](LICENSE).
 
 RAPTOR requires requires Python 3 (tested with Python 3.8+). The following Python packages are necessary:
 ```bash
-    numpy, numba, pyyaml, vtk, scikit-image, pandas
+    numpy, numba, pyyaml, vtk, scikit-image, pandas, pyvista
 ```
 
 *   **NumPy**: For numerical operations and array manipulation.
@@ -39,6 +39,7 @@ RAPTOR requires requires Python 3 (tested with Python 3.8+). The following Pytho
 *   **VTK**: For writing the output porosity map in `.vti` format
 *   **scikit-image**: For calculating pore morphologies.
 *   **pandas**: For writing morphology information to .csv
+*   **pyvista**: For visualization of `.vti` results.
 
 You can install all dependencies and Raptor itself by running ```pip install .``` in the cloned Raptor directory.
 
@@ -55,29 +56,29 @@ The project is organized into several modules:
 *   `io.py`: Contains functions for reading and parsing input files (scan paths, melt pool data).
 *   `utilities.py`: Includes helper classes, such as the `ScanPathBuilder` for generating scan strategies.
 
-RAPTOR can be used in two primary ways: through its Command-Line Interface (CLI) for quick, configuration-driven simulations, or as a Python Library (API) for integration into custom scripts and more complex workflows.
+RAPTOR can be used in two primary ways: through its Command-Line Interface (CLI) for quick, configuration-driven simulations, or as a Python Library (API) for integration into custom scripts and simulation workflows.
 
 ### 1. Command-Line Interface (CLI)
 
-The CLI is the simplest way to run a simulation. It is controlled by a single YAML configuration file that defines all inputs, parameters, and outputs.
+The CLI usage requires scan path files corresponding to build information. These scan path files can be generated with the `ScanPathBuilder` class in `raptor.utilities`. The CLI is controlled by a single YAML input file that defines all inputs, parameters, and outputs. The CLI example contains a single scan path to show functionality and observe the fluctuations of the simulated melt pool. The API example is recommended for a more descriptive simulation of undermelting-induced defects.
 
 **How to Run (CLI):**
 
-1.  **Prepare Inputs**: Create scan path files, melt pool data files, and a `config.yaml` file (detailed below).
+1.  **Prepare Inputs**: Create scan path files, melt pool data files, and a `input.yaml` file (detailed below).
 2.  **Execute Script**: Run the following command from your terminal, providing the path to your configuration file:
     ```bash
-    raptor path/to/your/config.yaml
+    raptor path/to/your/input.yaml
     ```
 3.  **Check Outputs**:
     *   Progress will be printed to the console.
     *   The 3D porosity map is saved to the `.vti` file specified in the config.
-    *   The pore morphology data is saved to the `.csv` file (if configured -- some CLI examples are set to not output morphology due to large domain bounding boxes).
+    *   The pore morphology data is saved to the `.csv` file (if configured -- the example does not save the morphology information.).
 
-#### CLI Input: The `config.yaml` File
+#### CLI Input: The `input.yaml` File
 
-Running RAPTOR from the CLI requires a YAML configuration file to specify all parameters.
+Running RAPTOR from the CLI requires a YAML input file to specify all parameters.
 
-**Example `config.yaml`:**
+**Example `input.yaml`:**
 ```yaml
 # List of scan path files (relative to this config file's location)
 scan_paths:
@@ -146,13 +147,15 @@ output:
    0 0.001 0.0001 0.000 200 0.8
    ```
 
[…]
