+++ examples/data/scanPathData/rve_layers/ULI_rve_layer_1.txt
@@ -1,17 +1,0 @@
-Mode X(m) Y(m) Z(m) Power(W) tParam
-1.000000 0.000515 -0.000406 0.000040 0.000000 0.000000
-0.000000 0.000906 0.000515 0.000040 370.000000 1.700000
-1.000000 0.000386 -0.000351 0.000040 0.000000 0.000000
-0.000000 0.000777 0.000570 0.000040 370.000000 1.700000
-1.000000 0.000257 -0.000296 0.000040 0.000000 0.000000
-0.000000 0.000648 0.000624 0.000040 370.000000 1.700000
-1.000000 0.000128 -0.000242 0.000040 0.000000 0.000000
-0.000000 0.000519 0.000679 0.000040 370.000000 1.700000
-1.000000 -0.000001 -0.000187 0.000040 0.000000 0.000000
-0.000000 0.000390 0.000734 0.000040 370.000000 1.700000
-1.000000 -0.000129 -0.000132 0.000040 0.000000 0.000000
-0.000000 0.000261 0.000788 0.000040 370.000000 1.700000
-1.000000 -0.000258 -0.000077 0.000040 0.000000 0.000000
-0.000000 0.000132 0.000843 0.000040 370.000000 1.700000
-1.000000 -0.000387 -0.000023 0.000040 0.000000 0.000000
-0.000000 0.000004 0.000898 0.000040 370.000000 1.700000

[…]
