### 2026-03-17 ef17187
revised maniscript
