+++ doc/paper.md
@@ -37,11 +37,11 @@ Metal AM processes show great promise in advancing manufacturing capabilities ac
 
 # State of the field
 
-The current modeling landscape for sLoF prediction is fairly sparse; initial work provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction, continuing to utilize the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model has been introduced [@richter_analyticallof_2025]. The current state of the art sLoF model builds on this approach with a Fourier-based expansion of the melt pool dimensions to account for temporal variability [@subraveti_sma_2025;@subraveti_process_2025]. This model solves for explicit melt pool overlaps, yielding sLoF geometries resultant from melt pool fluctuations. 
+The current modeling landscape for sLoF prediction is fairly sparse; initial work provided simple 2-dimensional estimations of defect volume fraction in the deterministic melt pool case [@tang_lof_2017]. This was extended to 3-dimensional geometry prediction, continuing to utilize the deterministic melt pool assumption [@subraveti_lof_2024]. More recently, uncertainty in the melt pool dimensions to augment sLoF predictions in a Tang-type semi-analytical geometric model has been introduced [@richter_analyticallof_2025]. The current state-of-the-art sLoF model builds on this approach with a Fourier-based expansion of the melt pool dimensions to account for temporal variability [@subraveti_sma_2025;@subraveti_process_2025]. This model solves for explicit melt pool overlaps, yielding sLoF geometries resultant from melt pool fluctuations. 
 
 # Software design
 
-In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. The novel parallelized algorithm we present results multiple order-of-magnitude efficiency gain over the previous state-of-the-art serially timestepped approach. In Raptor, the melt pool dimensions are treated as a stochastic process informed by user-supplied data. These stochastic processes are efficiently sampled via a truncated cosine expansion of the full Fourier basis computed from the user-supplied data. The locally computed dimensions from the cosine expansion are used to define a Lamé curve, which is then used to mask a voxelized grid representing the printed volume of interest.
+In this manuscript, we introduce a significant improvement to state-of-the-art in the prediction and characterization of explicit sLoF geometries and the distribution of their morphological metrics. The novel parallelized algorithm in Raptor results in a multiple order-of-magnitude efficiency gain over the previous state-of-the-art serially timestepped approach. In Raptor, the melt pool dimensions are treated as a stochastic process informed by user-supplied data. These stochastic processes are efficiently sampled via a truncated cosine expansion of the full Fourier basis computed from the user-supplied data. The locally computed dimensions from the cosine expansion are used to define a Lamé curve, which is then used to mask a voxelized grid representing the printed volume of interest.
 
[…]
