### 2026-03-17 93b6c27
cleaned readme and added contributor instructions.
