+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+Contributing to Raptor is easy: just open a pull request. Make main the destination branch on the Raptor repository and allow edits from maintainers.
+
+Your pull request must work with all current Raptor tutorial examples and be reviewed by at least one of the main developers.
+
+We use `pre-commit` to fix formatting into a consistent style. You can install `pre-commit` through `pip`.
