### 2026-03-17 9a56192
turning off pyvista rendering to pass CI
