+++ examples/api_example/rve.py
@@ -93,4 +93,4 @@ morphology = compute_morphology(
 write_morphology(morphology, "rve_morphology.csv")
 
 # 7. Visualize using PyVista
-visualize("./rve.vti")
+# visualize("./rve.vti")
