### 2026-03-18 25ab75c
restored original filenames in input.yaml
