### 2026-03-18 a2e3efc
Joss 2025 revisions (#25)
* added morphology writing to api. renamed scan path file in cli

* added complex cli example.

* updated some examples and readme.

* new rve example and updated readme.

* cleaned readme and added contributor instructions.

* updated integrated tests

* fixed input.yaml

* fixing plotting

* turning off pyvista rendering to pass CI

* updated readme

---------

Co-authored-by: Vamsi Subraveti <vis@ornl.gov>
