+++ pyproject.toml
@@ -26,7 +26,9 @@ dependencies = [
     "PyYAML",
     "vtk",
     "scikit-image",
-    "pytest"
+    "pytest",
+    "pandas",
+    "pyvista"
 ]
 
 [project.scripts]
