### 2026-03-20 b0c7b2b
C++ Port Plus Optimizations
