+++ C++/src/api.cpp
@@ -1,0 +1,282 @@
+#include "raptor/api.hpp"
+
+#include <Kokkos_Core.hpp>
+#include <Kokkos_Profiling_ScopedRegion.hpp>
+
+#include <algorithm>
+#include <array>
+#include <cmath>
+#include <complex>
+#include <map>
+#include <random>
+#include <stdexcept>
+#include <vector>
+
+namespace raptor {
+namespace {
+
+constexpr double kPi = 3.14159265358979323846;
+
+// Convert raw numeric rows into the fixed-width time-series view expected by the DFT helper.
+template <typename Real>
+std::vector<std::array<Real, 2>> toTimeSeries(const std::vector<std::vector<Real>>& rows) {
+  std::vector<std::array<Real, 2>> result;
+  result.reserve(rows.size());
+  for (const std::vector<Real>& row : rows) {
+    if (row.size() != 2) {
+      throw std::invalid_argument("Time-series data must have exactly two columns.");
+    }
+    result.push_back({row[0], row[1]});
+  }
+  return result;
+}
+
+// Rotate a 2D point around the hatch-center used to generate successive layers.
+template <typename Real>
+std::array<Real, 2> rotatePoint(const std::array<Real, 2>& point,
+                                const std::array<Real, 2>& center, Real angle) {
+  const Real cosine = static_cast<Real>(std::cos(angle));
+  const Real sine = static_cast<Real>(std::sin(angle));
+  const Real local_x = point[0] - center[0];
+  const Real local_y = point[1] - center[1];
+  return {cosine * local_x - sine * local_y + center[0],
+          sine * local_x + cosine * local_y + center[1]};
+}
+
+}  // namespace
+
+template <typename Real>
+GridT<Real> createGrid(Real voxel_resolution, const std::optional<BoundBoxT<Real>>& bound_box,
+                       const std::vector<PathVectorT<Real>>* path_vectors) {
+  // Preserve the Python API wrapper semantics around Grid construction.
+  return GridT<Real>(voxel_resolution, bound_box, path_vectors);
+}
+
+template <typename Real>
+std::vector<PathVectorT<Real>> createPathVectors(const BoundBoxT<Real>& bound_box, Real power,
+                                                 Real scan_speed, Real hatch_spacing,
+                                                 Real layer_height, Real rotation_degrees,
+                                                 Real scan_extension, int extra_layers) {
+  (void)power;
+
+  // Recreate the Python scan-path builder without introducing extra filtering behavior.
+  const Vec3T<Real> min_point = bound_box[0];
+  const Vec3T<Real> max_point = bound_box[1];
+  const Vec3T<Real> dimensions = {max_point[0] - min_point[0], max_point[1] - min_point[1],
+                                  max_point[2] - min_point[2]};
+  const std::array<Real, 2> center = {(min_point[0] + max_point[0]) / static_cast<Real>(2),
+                                      (min_point[1] + max_point[1]) / static_cast<Real>(2)};
+  const Real rotation =
+      rotation_degrees * static_cast<Real>(kPi / 180.0);
+  const int n_layers =
+      static_cast<int>(std::floor(dimensions[2] / layer_height) + 1.0) + extra_layers;
+
+  const Real xmin = min_point[0] - scan_extension;
+  const Real xmax = max_point[0] + scan_extension;
+  const Real ymin = min_point[1] - scan_extension;
+  const Real ymax = max_point[1] + scan_extension;
+
+  std::vector<std::array<Real, 2>> base_starts;
+  std::vector<std::array<Real, 2>> base_ends;
+  for (Real y = ymin; y < ymax; y += hatch_spacing) {
+    base_starts.push_back({xmin, y});
+    base_ends.push_back({xmax, y});
[…]
