+++ C++/tests/smoke.cpp
@@ -1,0 +1,300 @@
+#include "raptor/api.hpp"
+
+#include <Kokkos_Core.hpp>
+
+#include <cstdlib>
+#include <cstring>
+#include <cctype>
+#include <filesystem>
+#include <fstream>
+#include <iostream>
+#include <iterator>
+#include <map>
+#include <optional>
+#include <stdexcept>
+#include <string>
+#include <vector>
+
+namespace {
+
+void require(bool condition, const std::string& message) {
+  if (!condition) {
+    throw std::runtime_error(message);
+  }
+}
+
+// Decode a base64-encoded appended payload from the VTI writer.
+std::vector<std::uint8_t> decodeBase64(const std::string& encoded) {
+  auto decodeChar = [](char ch) -> int {
+    if (ch >= 'A' && ch <= 'Z') {
+      return ch - 'A';
+    }
+    if (ch >= 'a' && ch <= 'z') {
+      return ch - 'a' + 26;
+    }
+    if (ch >= '0' && ch <= '9') {
+      return ch - '0' + 52;
+    }
+    if (ch == '+') {
+      return 62;
+    }
+    if (ch == '/') {
+      return 63;
+    }
+    if (ch == '=') {
+      return -2;
+    }
+    return -1;
+  };
+
+  std::string filtered;
+  filtered.reserve(encoded.size());
+  for (char ch : encoded) {
+    if (!std::isspace(static_cast<unsigned char>(ch))) {
+      filtered.push_back(ch);
+    }
+  }
+
+  require(filtered.size() % 4 == 0, "Base64 payload length must be a multiple of 4.");
+
+  std::vector<std::uint8_t> decoded;
+  decoded.reserve((filtered.size() / 4) * 3);
+
+  for (std::size_t index = 0; index < filtered.size(); index += 4) {
+    const int c0 = decodeChar(filtered[index]);
+    const int c1 = decodeChar(filtered[index + 1]);
+    const int c2 = decodeChar(filtered[index + 2]);
+    const int c3 = decodeChar(filtered[index + 3]);
+    require(c0 >= 0 && c1 >= 0, "Invalid base64 prefix.");
+    require(c2 != -1 && c3 != -1, "Invalid base64 suffix.");
+
+    const std::uint32_t triple = (static_cast<std::uint32_t>(c0) << 18) |
+                                 (static_cast<std::uint32_t>(c1) << 12) |
+                                 (static_cast<std::uint32_t>(c2 >= 0 ? c2 : 0) << 6) |
+                                 static_cast<std::uint32_t>(c3 >= 0 ? c3 : 0);
+    decoded.push_back(static_cast<std::uint8_t>((triple >> 16) & 0xFFU));
+    if (c2 != -2) {
+      decoded.push_back(static_cast<std::uint8_t>((triple >> 8) & 0xFFU));
+    }
+    if (c3 != -2) {
+      decoded.push_back(static_cast<std::uint8_t>(triple & 0xFFU));
+    }
+  }
+
+  return decoded;
+}
+
+void testSpectralModeBounds() {
+  // Verify that the C++ port now rejects more requested modes than available samples.
+  const std::vector<std::array<double, 2>> samples = {{0.0, 1.0}, {1.0, 2.0}, {2.0, 3.0}};
+  bool threw = false;
+  try {
+    (void)raptor::computeSpectralComponents(samples, 4);
+  } catch (const std::invalid_argument&) {
+    threw = true;
+  }
+  require(threw, "computeSpectralComponents should reject n_modes > sample_count.");
+}
+
+void testMorphologyCentroidAndBBox() {
+  // Verify that morphology centroids are computed from voxel centers and bbox is preserved.
[…]
