### 2026-03-21 b0c13b0
Fast Teams Bitwise Packed
