+++ C++/src/core.cpp
@@ -14,6 +14,8 @@ namespace raptor {
 namespace {
 
 constexpr int kMaxSeedBatchLanes = 4;
+constexpr int kPackedMaskWordBits = 64;
+constexpr int kTeamRepeat64Lanes = 64;
 constexpr int kTeamTileVoxels = 32;
 constexpr int kTeamVectorChunk = 16;
 
@@ -58,14 +60,6 @@ KOKKOS_INLINE_FUNCTION Real clamp01(Real value) {
              : (value > static_cast<Real>(1) ? static_cast<Real>(1) : value);
 }
 
-// Resolve the runtime selector into the actual kernel to launch.
-PorosityKernelVariant resolvePorosityVariant(PorosityKernelVariant requested) {
-  if (requested == PorosityKernelVariant::auto_select) {
-    return PorosityKernelVariant::baseline;
-  }
-  return requested;
-}
-
 // Check whether every lane in a batched-seed kernel has already melted.
 KOKKOS_INLINE_FUNCTION bool allLanesMelted(const std::uint8_t lane_melted[],
                                            int lane_count) {
@@ -75,6 +69,20 @@ KOKKOS_INLINE_FUNCTION bool allLanesMelted(const std::uint8_t lane_melted[],
     }
   }
   return true;
+}
+
+// Build the all-ones mask corresponding to the active repeat bits in one packed batch.
+template <typename MaskType>
+KOKKOS_INLINE_FUNCTION MaskType fullMaskForBits(int bit_count) {
+  constexpr int kMaskBits = static_cast<int>(sizeof(MaskType) * 8);
+  return bit_count >= kMaskBits ? static_cast<MaskType>(~static_cast<MaskType>(0))
+                                : static_cast<MaskType>((static_cast<MaskType>(1) << bit_count) -
+                                                        static_cast<MaskType>(1));
+}
+
+// Convert a repeat count into the number of packed 64-bit mask words needed per voxel.
+KOKKOS_INLINE_FUNCTION int packedMaskWordCount(int repeat_count) {
+  return (repeat_count + kPackedMaskWordBits - 1) / kPackedMaskWordBits;
 }
 
 // Fill the packed path-vector row with geometry and melt-pool bounds that are stable across runs.
@@ -187,6 +195,36 @@ KOKKOS_INLINE_FUNCTION Real phaseFromSeed(std::uint64_t seed, int vector, int mo
   return static_cast<Real>(2.0 * 3.14159265358979323846 * unit);
 }
 
+// Route float CUDA builds through NVIDIA fast trig intrinsics while preserving the generic path
+// for double precision and non-CUDA backends.
+template <typename Real>
+KOKKOS_INLINE_FUNCTION Real fastCos(Real angle) {
+  return Kokkos::cos(angle);
+}
+
+KOKKOS_INLINE_FUNCTION float fastCos(float angle) {
+#if defined(__CUDA_ARCH__)
+  return __cosf(angle);
+#else
+  return Kokkos::cos(angle);
+#endif
+}
+
+template <typename Real>
+KOKKOS_INLINE_FUNCTION void fastSinCos(Real angle, Real& sine, Real& cosine) {
+  sine = Kokkos::sin(angle);
+  cosine = Kokkos::cos(angle);
+}
+
+KOKKOS_INLINE_FUNCTION void fastSinCos(float angle, float& sine, float& cosine) {
+#if defined(__CUDA_ARCH__)
+  __sincosf(angle, &sine, &cosine);
+#else
+  sine = Kokkos::sin(angle);
+  cosine = Kokkos::cos(angle);
+#endif
+}
+
 // Fill the phase matrix on device so repeat runs only change the seed and reuse static buffers.
[…]
