+++ C++/apps/RVE-teamBench.cpp
@@ -1,0 +1,166 @@
+#include "raptor/api.hpp"
+
+#include <Kokkos_Core.hpp>
+#include <Kokkos_Profiling_ScopedRegion.hpp>
+
+#include <array>
+#include <cstdlib>
+#include <filesystem>
+#include <iostream>
+#include <map>
+#include <stdexcept>
+#include <string>
+#include <vector>
+
+namespace raptor {
+namespace {
+
+// Report the active Kokkos execution space so benchmark logs show the backend used.
+void printExecutionSpace() {
+  std::cout << "Kokkos execution space: " << Kokkos::DefaultExecutionSpace::name() << '\n';
+}
+
+// Round a positive integer down to the nearest power of two.
+int floorPowerOfTwo(int value) {
+  int power = 1;
+  while ((power << 1) > 0 && (power << 1) <= value) {
+    power <<= 1;
+  }
+  return power;
+}
+
+// Read an unsigned integer benchmark override from the environment.
+std::uint64_t readEnvUint64(const char* name, std::uint64_t fallback) {
+  const char* value = std::getenv(name);
+  if (value == nullptr) {
+    return fallback;
+  }
+  return static_cast<std::uint64_t>(std::stoull(value));
+}
+
+// Build the power-of-two team-size sweep from the active backend and Kokkos-reported bounds.
+std::vector<int> buildTeamSizes() {
+  const std::string execution_space = Kokkos::DefaultExecutionSpace::name();
+  const TeamBitpackedSizeBounds bounds = queryTeamBitpackedSizeBounds<float>();
+
+  int min_team_size = 1;
+  int max_team_size = floorPowerOfTwo(std::max(1, bounds.maximum));
+  if (execution_space == "Cuda" || execution_space == "HIP" || execution_space == "SYCL" ||
+      execution_space == "OpenMPTarget") {
+    min_team_size = 32;
+  }
+
+  std::vector<int> team_sizes;
+  for (int team_size = min_team_size; team_size <= max_team_size; team_size <<= 1) {
+    team_sizes.push_back(team_size);
+  }
+  return team_sizes;
+}
+
+// Hold the reusable float-precision RVE inputs for the team-size sweep benchmark.
+struct TeamBenchCaseData {
+  GridT<float> grid;
+  std::vector<PathVectorT<float>> path_vectors;
+  MeltPoolT<float> melt_pool;
+};
+
+// Build the fixed float RVE benchmark inputs once so only the kernel strategy varies.
+TeamBenchCaseData buildBenchmarkCase() {
+  TeamBenchCaseData data;
+
+  const std::array<float, 3> min_point = {0.0f, 0.0f, 0.0f};
+  const std::array<float, 3> max_point = {5.0e-4f, 5.0e-4f, 5.0e-4f};
+  const BoundBoxT<float> bound_box = {min_point, max_point};
+  {
+    Kokkos::Profiling::ScopedRegion region("raptor::RVETeamBench::create_grid");
+    data.grid = createGrid<float>(5.0e-6f, std::optional<BoundBoxT<float>>(bound_box), nullptr);
+  }
+
+  {
+    Kokkos::Profiling::ScopedRegion region("raptor::RVETeamBench::create_path_vectors");
+    data.path_vectors = createPathVectors<float>(bound_box, 370.0f, 1.7f, 140.0e-6f, 30.0e-6f,
+                                                 67.0f, 5.0e-4f, 0);
+  }
+
+  const std::filesystem::path melt_pool_data_path =
+      std::filesystem::current_path() / ".." / "data" / "meltPoolData" /
+      "ULI_v1700_theta0_widths.txt";
+  const std::vector<std::vector<float>> width_data = readData<float>(melt_pool_data_path);
[…]
