### 2026-04-10 57eedb6
use dial for acive learning
