+++ examples/api_example/active_learning_dial.py
@@ -1,0 +1,457 @@
+import argparse
+import json
+import logging
+import os
+import sys
+import random
+from pathlib import Path
+from typing import Any
+
+import numpy as np
+import matplotlib
+
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+from scipy.stats import qmc
+
+# Raptor Imports
+from raptor.io import read_data
+from raptor.api import (
+    create_grid,
+    create_melt_pool,
+    compute_porosity,
+    compute_morphology,
+)
+from raptor.utilities import ScanPathBuilder
+
+# Intersect Imports
+from intersect_sdk import (
+    INTERSECT_JSON_VALUE,
+    HierarchyConfig,
+    IntersectClient,
+    IntersectClientCallback,
+    IntersectClientConfig,
+    IntersectDirectMessageParams,
+    default_intersect_lifecycle_loop,
+)
+
+# Dial Imports
+from dial_dataclass import (
+    DialInputPredictions,
+    DialInputSingleOtherStrategy,
+    DialWorkflowCreationParamsClient,
+    DialWorkflowDatasetUpdate,
+)
+
+
+# Logging
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s %(levelname)-8s %(message)s"
+)
+logger = logging.getLogger(__name__)
+
+
+# -----------------------------------------------------------------------------
+# USER PARAMETERS
+# -----------------------------------------------------------------------------
+INITIAL_BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
+UNIT_BOUNDS = [[0.0, 1.0]]
+NUM_DIMS = len(INITIAL_BOUNDS)
+
+INITIAL_DATA_SIZE = 8   # Number of initial LHS points
+MAX_ITERATIONS = 20     # Number of active learning iterations
+MESHGRID_SIZE = 100
+
+# Ensemble parameters (sampling around suggested point)
+ENSEMBLE_SIZE = 8
+LOCAL_BOX_FRAC = 0.1
+
+INITIAL_POINTS_TO_PREDICT = np.linspace(
+    INITIAL_BOUNDS[0][0],
+    INITIAL_BOUNDS[0][1],
+    MESHGRID_SIZE
+).reshape(-1, 1)
+
+# History tracking
+HISTORY_X: list[list[float]] = []
+HISTORY_Y: list[float] = []
+
+
+# -----------------------------------------------------------------------------
+# RAPTOR
+# -----------------------------------------------------------------------------
+def compute_mean_and_std(defects: np.ndarray):
+    if len(defects) == 0:
+        y = 0.0
+        yerr = np.nan
+    elif len(defects) == 1:
+        y = defects[0]
+        yerr = defects[0]
+    else:
+        y = np.mean(defects)
+        yerr = np.std(defects, ddof=1) / np.sqrt(len(defects))
+
+    return y, yerr
+
+
+def run_raptor_step(
+    hatch_spacing_m: float,
+    layer_thickness_m: float = 50e-6,
+    query_volume_mm3: float = 1.0,
+    voxel_resolution_m: float = 2.5e-6,
+    metric_names: list[str] = ["equivalent_diameter_area"],   
+    ):
+    """Executes a single Raptor simulation."""
+    rve_min_point = np.array([0.0, 0.0, 0.0])
+    rve_max_point = np.array([5e-4, 5e-4, 5e-4])
+    rve_bounding_box = np.array([rve_min_point, rve_max_point])
+
+    grid = create_grid(
+        voxel_resolution=voxel_resolution_m,
+        bound_box=rve_bounding_box
+    )
+
+    SCRIPT_DIR = Path(__file__).resolve().parent
+    melt_pool_data_path = (
[…]
