+++ pyproject.toml
@@ -4,13 +4,13 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "raptor"
-version = "0.1.0"
+version = "1.0.0"
 license = { text = "BSD-3-Clause" }
 authors = [
   { name="John Coleman", email="colemanjs@ornl.gov" },
   { name="Vamsi Subraveti", email="vamsi.r.subraveti@vanderbilt.edu" }
 ]
-description = "A library to predict porosity in Laser Powder Bed Fusion (LPBF) builds."
+description = "A library to predict porosity in Additive Manufacturing (AM)."
 readme = "README.md"
 requires-python = ">=3.8"
 classifiers = [
@@ -26,11 +26,15 @@ dependencies = [
     "PyYAML",
     "vtk",
     "scikit-image",
-    "pytest"
+    "pytest",
+    "pandas",
+    "pyvista",
+    "intersect-sdk",
+    "dial @ git+https://github.com/INTERSECT-DIAL/dial"
 ]
 
 [project.scripts]
 raptor = "raptor.cli:run"
 
 [project.urls]
-Homepage = "https://code.ornl.gov/ORNL-MDF/rapiddefectpredictor"
+Homepage = "https://github.com/ORNL-MDF/Raptor"
