### 2026-04-13 ffe1eaa
fixup: smaller example for active learning
