+++ examples/api_example/active_learning_dial.py
@@ -54,21 +54,21 @@ logger = logging.getLogger(__name__)
 # -----------------------------------------------------------------------------
 # USER PARAMETERS
 # -----------------------------------------------------------------------------
-INITIAL_BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
+BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
 UNIT_BOUNDS = [[0.0, 1.0]]
-NUM_DIMS = len(INITIAL_BOUNDS)
+NUM_DIMS = len(BOUNDS)
 
 INITIAL_DATA_SIZE = 8   # Number of initial LHS points
 MAX_ITERATIONS = 20     # Number of active learning iterations
-MESHGRID_SIZE = 100
+MESHGRID_SIZE = 101
 
 # Ensemble parameters (sampling around suggested point)
 ENSEMBLE_SIZE = 8
-LOCAL_BOX_FRAC = 0.1
+LOCAL_FRAC = 0.1
 
 INITIAL_POINTS_TO_PREDICT = np.linspace(
-    INITIAL_BOUNDS[0][0],
-    INITIAL_BOUNDS[0][1],
+    BOUNDS[0][0],
+    BOUNDS[0][1],
     MESHGRID_SIZE
 ).reshape(-1, 1)
 
@@ -94,13 +94,13 @@ def compute_mean_and_std(defects: np.ndarray):
     return y, yerr
 
 
-def run_raptor_step(
+def run_raptor(
     hatch_spacing_m: float,
     layer_thickness_m: float = 50e-6,
     query_volume_mm3: float = 1.0,
-    voxel_resolution_m: float = 2.5e-6,
-    metric_names: list[str] = ["equivalent_diameter_area"],   
-    ):
+    voxel_resolution_m: float = 5.0e-6,
+    metric_names: list[str] = ["equivalent_diameter_area"],
+):
     """Executes a single Raptor simulation."""
     rve_min_point = np.array([0.0, 0.0, 0.0])
     rve_max_point = np.array([5e-4, 5e-4, 5e-4])
@@ -161,14 +161,14 @@ def run_raptor_step(
     # Run Raptor and calculate defect metrics
     rve_volume_mm3 = (1e3 * (rve_max_point - rve_min_point)).prod()
     num_rves = int(np.ceil(query_volume_mm3 / rve_volume_mm3))
-    
+
     defect_metrics = []
     for i in range(num_rves):
         porosity = compute_porosity(
             grid,
             path_vectors,
             melt_pool,
-            (i == 0)  # jit warmup
+            False  # jit warmup
         )
         metrics = compute_morphology(porosity, grid.resolution, metric_names)
         defect_metrics.append(metrics)
@@ -183,45 +183,47 @@ def run_raptor_step(
             combined_metrics[name] = np.array([])
 
     mean, _ = compute_mean_and_std(combined_metrics["equivalent_diameter_area"])
-    
-    print("Hatch spacing: ", hatch_spacing, "Mean pore size: ", mean)
+
+    print("Hatch spacing: ", hatch_spacing_m, "Mean pore size: ", mean)
     return mean
 
 
 # -----------------------------------------------------------------------------
 # UTILITIES
 # -----------------------------------------------------------------------------
-def x_to_unit_arr(X):
+def x_to_unit(X):
     X = np.asarray(X, dtype=float)
-    lo = np.array([b[0] for b in INITIAL_BOUNDS])
-    hi = np.array([b[1] for b in INITIAL_BOUNDS])
+    lo = np.array([b[0] for b in BOUNDS])
[…]
