### 2026-04-13 1f5990a
removed random sampling around suggested point
