+++ examples/api_example/active_learning_dial.py
@@ -58,23 +58,19 @@ BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
 UNIT_BOUNDS = [[0.0, 1.0]]
 NUM_DIMS = len(BOUNDS)
 
-INITIAL_DATA_SIZE = 8   # Number of initial LHS points
-MAX_ITERATIONS = 20     # Number of active learning iterations
+INITIAL_DATA_SIZE = 8  # Number of initial LHS points
+MAX_ITERATIONS = 25  # Number of active learning (AL) iterations
+
+# Global seed for deterministic active learning
+SEED = 42
+random.seed(SEED)
+np.random.seed(SEED)
+
 MESHGRID_SIZE = 101
 
-# Ensemble parameters (sampling around suggested point)
-ENSEMBLE_SIZE = 8
-LOCAL_FRAC = 0.1
-
 INITIAL_POINTS_TO_PREDICT = np.linspace(
-    BOUNDS[0][0],
-    BOUNDS[0][1],
-    MESHGRID_SIZE
+    BOUNDS[0][0], BOUNDS[0][1], MESHGRID_SIZE
 ).reshape(-1, 1)
-
-# History tracking
-HISTORY_X: list[list[float]] = []
-HISTORY_Y: list[float] = []
 
 
 # -----------------------------------------------------------------------------
@@ -106,10 +102,7 @@ def run_raptor(
     rve_max_point = np.array([5e-4, 5e-4, 5e-4])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
 
-    grid = create_grid(
-        voxel_resolution=voxel_resolution_m,
-        bound_box=rve_bounding_box
-    )
+    grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
 
     SCRIPT_DIR = Path(__file__).resolve().parent
     melt_pool_data_path = (
@@ -164,12 +157,7 @@ def run_raptor(
 
     defect_metrics = []
     for i in range(num_rves):
-        porosity = compute_porosity(
-            grid,
-            path_vectors,
-            melt_pool,
-            False  # jit warmup
-        )
+        porosity = compute_porosity(grid, path_vectors, melt_pool, False)  # jit warmup
         metrics = compute_morphology(porosity, grid.resolution, metric_names)
         defect_metrics.append(metrics)
 
@@ -182,9 +170,11 @@ def run_raptor(
         else:
             combined_metrics[name] = np.array([])
 
-    mean, _ = compute_mean_and_std(combined_metrics["equivalent_diameter_area"])
+    mean, std = compute_mean_and_std(combined_metrics["equivalent_diameter_area"])
 
-    print("Hatch spacing: ", hatch_spacing_m, "Mean pore size: ", mean)
+    print("Hatch spacing: ", hatch_spacing_m)
+    print("Pore size (mean): ", mean)
+    print("Pore size (std):  ", std)
     return mean
 
 
@@ -207,30 +197,12 @@ def x_from_unit(U):
 
 def get_data_point(x_suggested):
     """
-    Runs the suggested point plus a few local variations to account for simulation noise.
+    Runs the suggested point.
     """
-    ensemble_xs = [x_suggested]
+    x_ = [np.clip(x_suggested[0], BOUNDS[0][0], BOUNDS[0][1])]
+    y_ = run_raptor(x_[0])
 
-    span = BOUNDS[0][1] - BOUNDS[0][0]
-    for _ in range(ENSEMBLE_SIZE - 1):
-        delta = (random.random() - 0.5) * 2 * LOCAL_FRAC * span
-        new_x = np.clip(
-            x_suggested[0] + delta,
[…]
