### 2026-04-14 0ba012b
added convolution filter
