+++ examples/active_learning_example/melt_pool_model/create_melt_pool_model.py
@@ -1,0 +1,320 @@
+import argparse
+import json
+import logging
+import os
+import sys
+from pathlib import Path
+from typing import Any
+
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+import pandas as pd
+
+mpl.use("agg")
+
+import numpy as np
+from intersect_sdk import (
+    INTERSECT_JSON_VALUE,
+    HierarchyConfig,
+    IntersectClient,
+    IntersectClientCallback,
+    IntersectClientConfig,
+    IntersectDirectMessageParams,
+    default_intersect_lifecycle_loop,
+)
+
+# from scipy.stats import qmc
+from dial_dataclass import (
+    DialInputPredictions,
+    DialInputSingleOtherStrategy,
+    DialWorkflowCreationParamsClient,
+    DialWorkflowDatasetUpdate,
+)
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+# AM SPECIFIC FUNCTIONS ------------------------------------------------------------------------------------------------------
+def am_dataset_coleman():
+    # List of JSON file names
+    json_files = [
+        "parameters_A.json",
+        "parameters_B.json",
+        "parameters_C.json",
+        "parameters_D.json",
+        "parameters_E.json",
+    ]
+
+    # Initialize a set to store all keys
+    all_keys = set()
+
+    # Loop through the JSON files to determine all keys
+    for json_file in json_files:
+        with open(json_file, "r") as file:
+            data = json.load(file)
+            all_keys.update(data.keys())
+
+    # Initialize a dictionary to store lists of data
+    data_dict = {key: [] for key in all_keys}
+    file_labels = []
+
+    # Loop through the JSON files to collect data and labels
+    for json_file in json_files:
+        with open(json_file, "r") as file:
+            data = json.load(file)
+            for key in all_keys:
+                data_dict[key].extend(data.get(key, [None] * len(data_dict[key])))
+            file_labels.extend(
+                [json_file.split("_")[1].split(".")[0]]
+                * len(data.get(list(data.keys())[0], []))
+            )
+
+    # Create a DataFrame using the collected data and labels
+    df = pd.DataFrame(data_dict)
+    df["Location"] = file_labels
+
+    df["average_depth"] = (df["right_depth"] + df["left_depth"]) / 2.0
+
+    dataset_x = [df["Power"], df["Velocity"]]
+    dataset_x = np.transpose(dataset_x)
+    print(dataset_x)
+    dataset_y = df["average_depth"]
+    print(dataset_y)
+
+    return dataset_x, dataset_y
+
+
+def graph(mean_grid, variance, dataset_x, dataset_y):
+    if NUM_DIMS == 2:
+        plt.clf()
+        plt.contourf(
+            INITIAL_MESHGRIDS[0],
+            INITIAL_MESHGRIDS[1],
+            mean_grid,
+            extend="both",
+        )
+        cbar = plt.colorbar()
+        cbar.set_label("Melt pool depth (um)")
+        plt.xlabel("Power (W)")
+        plt.ylabel("Velocity (um/s)")
+        # add black dots for data points and a red marker for the recommendation:
+        X_train = np.array(dataset_x)
+        plt.scatter(
+            X_train[:, 0], X_train[:, 1], facecolor="none", color="black", marker="o"
+        )
+
+        plt.savefig("function_value.png")
+
+        plt.clf()
+        plt.contourf(
+            INITIAL_MESHGRIDS[0],
+            INITIAL_MESHGRIDS[1],
+            variance,
+            extend="both",
+        )
+        cbar = plt.colorbar()
+        cbar.set_label("Melt pool depth variance (um)")
[…]
