### 2026-04-14 3826cc8
added melt pool surrogate to active learning example
