### 2026-04-15 ed13979
input scaling for melt pool model
