+++ examples/active_learning_example/melt_pool_model/create_melt_pool_model.py
@@ -23,7 +23,6 @@ from intersect_sdk import (
     default_intersect_lifecycle_loop,
 )
 
-# from scipy.stats import qmc
 from dial_dataclass import (
     DialInputPredictions,
     DialInputSingleOtherStrategy,
@@ -39,6 +38,7 @@ logger = logging.getLogger(__name__)
 # USER PARAMETERS
 # -----------------------------------------------------------------------------
 INITIAL_BOUNDS = [[500, 1500], [100, 300]]  # [velocity, power]
+UNIT_BOUNDS = [[0.0, 1.0], [0.0, 1.0]]
 NUM_DIMS = len(INITIAL_BOUNDS)
 MESHGRID_SIZE = 101
 INITIAL_MESHGRIDS = np.meshgrid(
@@ -49,14 +49,26 @@ INITIAL_MESHGRIDS = np.meshgrid(
     indexing="ij",
 )
 INITIAL_POINTS_TO_PREDICT = np.hstack([mg.reshape(-1, 1) for mg in INITIAL_MESHGRIDS])
-NUM_ITERATIONS = 35
 
 
 # -----------------------------------------------------------------------------
 # UTILITIES
 # -----------------------------------------------------------------------------
+def x_to_unit(X):
+    X = np.asarray(X, dtype=float)
+    lo = np.array([b[0] for b in INITIAL_BOUNDS])
+    hi = np.array([b[1] for b in INITIAL_BOUNDS])
+    return (X - lo) / (hi - lo + 1e-12)
+
+
+def x_from_unit(U):
+    U = np.asarray(U, dtype=float)
+    lo = np.array([b[0] for b in INITIAL_BOUNDS])
+    hi = np.array([b[1] for b in INITIAL_BOUNDS])
+    return U * (hi - lo) + lo
+
+
 def melt_pool_dataset():
-    # List of JSON file names
     json_files = [
         "parameters_A.json",
         "parameters_B.json",
@@ -68,39 +80,76 @@ def melt_pool_dataset():
     all_keys = set()
 
     for json_file in json_files:
-        with open(json_file, "r") as file:
-            data = json.load(file)
-            all_keys.update(data.keys())
+        try:
+            with open(json_file, "r") as file:
+                data = json.load(file)
+                all_keys.update(data.keys())
+        except FileNotFoundError:
+            pass
 
     data_dict = {key: [] for key in all_keys}
-    file_labels = []
 
     for json_file in json_files:
-        with open(json_file, "r") as file:
-            data = json.load(file)
-            for key in all_keys:
-                data_dict[key].extend(data.get(key, [None] * len(data_dict[key])))
-            file_labels.extend(
-                [json_file.split("_")[1].split(".")[0]]
-                * len(data.get(list(data.keys())[0], []))
-            )
+        try:
+            with open(json_file, "r") as file:
+                data = json.load(file)
+                for key in all_keys:
+                    data_dict[key].extend(
+                        data.get(key, [None] * len(next(iter(data.values()))))
+                    )
+        except FileNotFoundError:
+            continue
 
     df = pd.DataFrame(data_dict)
-    df["Location"] = file_labels
 
     power = pd.concat([df["Power"], df["Power"]], ignore_index=True)
     velocity = pd.concat([df["Velocity"], df["Velocity"]], ignore_index=True)
 
-    dataset_x = np.stack([velocity, power], axis=1)
-    dataset_y = pd.concat([df["right_depth"], df["left_depth"]], ignore_index=True)
+    depth = pd.concat([df["right_depth"], df["left_depth"]], ignore_index=True)
+    width = pd.concat([df["right_width"], df["left_width"]], ignore_index=True)
[…]
