### 2026-04-15 433d118
use melt pool surrogate to run raptor at process set point
