+++ examples/active_learning_example/defect_model/create_defect_model_1D.py
@@ -13,7 +13,6 @@ import matplotlib
 matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 from scipy.stats import qmc
-from scipy.interpolate import RegularGridInterpolator
 
 # Raptor Imports
 from numba import njit
@@ -56,8 +55,8 @@ logger = logging.getLogger(__name__)
 # -----------------------------------------------------------------------------
 # USER PARAMETERS
 # -----------------------------------------------------------------------------
-LASER_POWER_WATTS = 200.0
-LASER_VELOCITY_M_S = 1.0
+WIDTH_MEAN = 148.0e-6
+WIDTH_STD = 18.0e-6
 
 BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
 UNIT_BOUNDS = [[0.0, 1.0]]
@@ -80,42 +79,10 @@ INITIAL_POINTS_TO_PREDICT = np.linspace(
     BOUNDS[0][0], BOUNDS[0][1], MESHGRID_SIZE
 ).reshape(-1, 1)
 
-MELT_POOL_SURROGATE_PATH = os.path.abspath(
-    os.path.join(
-        os.path.dirname(__file__), "..", "melt_pool_model", "melt_pool_surrogates.npz"
-    )
-)
-
 
 # -----------------------------------------------------------------------------
-# RAPTOR UTILITIES
+# RAPTOR
 # -----------------------------------------------------------------------------
-class MeltPoolInterpolator:
-    def __init__(self, filepath: str):
-        data = np.load(filepath)
-        self.v_axis = data["velocity"]
-        self.p_axis = data["power"]
-
-        self.features = [
-            "depth_mean",
-            "depth_std",
-            "width_mean",
-            "width_std",
-            "height_mean",
-            "height_std",
-        ]
-
-        self.interpolators = {}
-        for f in self.features:
-            self.interpolators[f] = RegularGridInterpolator(
-                (self.v_axis, self.p_axis), data[f]
-            )
-
-    def query(self, velocity: float, power: float):
-        point = np.array([velocity, power])
-        return {f: float(self.interpolators[f](point)) for f in self.features}
-
-
 def compute_mean_and_std(defects: np.ndarray):
     if len(defects) == 0:
         y = 0.0
@@ -132,79 +99,80 @@ def compute_mean_and_std(defects: np.ndarray):
 
 def run_raptor(
     hatch_spacing_m: float,
-    mp_interpolator: MeltPoolInterpolator,
-    layer_thickness_m: float = 40e-6,
+    layer_thickness_m: float = 50e-6,
     query_volume_mm3: float = 1.0,
     voxel_resolution_m: float = 5.0e-6,
     metric_names: list[str] = ["equivalent_diameter_area"],
 ):
-    """Executes a single Raptor simulation using dynamic melt pool surrogates."""
+    """Executes a single Raptor simulation."""
     # Re-seed at the start of every ensemble for independent determinism
     random.seed(SEED)
     np.random.seed(SEED)
     seed_numba(SEED)
 
-    # 1. Query the surrogates for power and velocity combination
-    mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
-
-    # 2. Create voxel grid
+    # 1. Create voxel grid
     rve_min_point = np.array([0.0, 0.0, 0.0])
     rve_max_point = np.array([5e-4, 5e-4, 5e-4])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
 
     grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
 
-    # 3. Build the laser scan path
+    # 2. Build the laser scan path
[…]
