### 2026-04-17 7d3329a
updated melt pool interpolator and defect criteria
