+++ examples/active_learning_example/defect_model/create_defect_model_1D.py
@@ -1,430 +1,0 @@
-import argparse
-import json
-import logging
-import os
-import sys
-import random
-from pathlib import Path
-from typing import Any
-
-import numpy as np
-import matplotlib
-
-matplotlib.use("Agg")
-import matplotlib.pyplot as plt
-from scipy.stats import qmc
-
-# Raptor Imports
-from numba import njit
-from raptor.io import read_data
-from raptor.api import (
-    create_grid,
-    create_melt_pool,
-    compute_porosity,
-    compute_morphology,
-)
-from raptor.utilities import ScanPathBuilder, MeltPoolFilter
-
-# Intersect Imports
-from intersect_sdk import (
-    INTERSECT_JSON_VALUE,
-    HierarchyConfig,
-    IntersectClient,
-    IntersectClientCallback,
-    IntersectClientConfig,
-    IntersectDirectMessageParams,
-    default_intersect_lifecycle_loop,
-)
-
-# Dial Imports
-from dial_dataclass import (
-    DialInputPredictions,
-    DialInputSingleOtherStrategy,
-    DialWorkflowCreationParamsClient,
-    DialWorkflowDatasetUpdate,
-)
-
-
-# Logging
-logging.basicConfig(
-    level=logging.INFO, format="%(asctime)s %(levelname)-8s %(message)s"
-)
-logger = logging.getLogger(__name__)
-
-
-# -----------------------------------------------------------------------------
-# USER PARAMETERS
-# -----------------------------------------------------------------------------
-WIDTH_MEAN = 148.0e-6
-WIDTH_STD = 18.0e-6
-
-BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
-UNIT_BOUNDS = [[0.0, 1.0]]
-NUM_DIMS = len(BOUNDS)
-
-INITIAL_DATA_SIZE = 8  # Number of initial LHS points
-MAX_ITERATIONS = 25  # Number of active learning (AL) iterations
-
-SEED = 42
-
-
-@njit
-def seed_numba(seed):
-    np.random.seed(seed)
-
-
-MESHGRID_SIZE = 101
-
-INITIAL_POINTS_TO_PREDICT = np.linspace(
-    BOUNDS[0][0], BOUNDS[0][1], MESHGRID_SIZE
-).reshape(-1, 1)
-
-
-# -----------------------------------------------------------------------------
-# RAPTOR
-# -----------------------------------------------------------------------------
-def compute_mean_and_std(defects: np.ndarray):
-    if len(defects) == 0:
-        y = 0.0
-        yerr = np.nan
-    elif len(defects) == 1:
-        y = defects[0]
-        yerr = defects[0]
-    else:
-        y = np.mean(defects)
-        yerr = np.std(defects, ddof=1) / np.sqrt(len(defects))
-
-    return y, yerr
-
-
-def run_raptor(
-    hatch_spacing_m: float,
-    layer_thickness_m: float = 50e-6,
-    query_volume_mm3: float = 1.0,
-    voxel_resolution_m: float = 5.0e-6,
-    metric_names: list[str] = ["equivalent_diameter_area"],
-):
-    """Executes a single Raptor simulation."""
-    # Re-seed at the start of every ensemble for independent determinism
-    random.seed(SEED)
-    np.random.seed(SEED)
-    seed_numba(SEED)
-
-    # 1. Create voxel grid
-    rve_min_point = np.array([0.0, 0.0, 0.0])
-    rve_max_point = np.array([5e-4, 5e-4, 5e-4])
-    rve_bounding_box = np.array([rve_min_point, rve_max_point])
-
[…]
