+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -8,10 +8,6 @@ from pathlib import Path
 from typing import Any
 
 import numpy as np
-import matplotlib
-
-matplotlib.use("Agg")
-import matplotlib.pyplot as plt
 from scipy.stats import qmc
 from scipy.interpolate import RegularGridInterpolator
 
@@ -59,12 +55,12 @@ logger = logging.getLogger(__name__)
 LASER_POWER_WATTS = 200.0
 LASER_VELOCITY_M_S = 1.0
 
-BOUNDS = [[60e-6, 200e-6]]  # Hatch spacing in meters
+BOUNDS = [[70e-6, 170e-6]]
 UNIT_BOUNDS = [[0.0, 1.0]]
 NUM_DIMS = len(BOUNDS)
 
-INITIAL_DATA_SIZE = 8  # Number of initial LHS points
-MAX_ITERATIONS = 25  # Number of active learning (AL) iterations
+INITIAL_DATA_SIZE = 20
+MAX_ITERATIONS = 35
 
 SEED = 42
 
@@ -74,7 +70,7 @@ def seed_numba(seed):
     np.random.seed(seed)
 
 
-MESHGRID_SIZE = 101
+MESHGRID_SIZE = 150
 
 INITIAL_POINTS_TO_PREDICT = np.linspace(
     BOUNDS[0][0], BOUNDS[0][1], MESHGRID_SIZE
@@ -112,22 +108,8 @@ class MeltPoolInterpolator:
             )
 
     def query(self, velocity: float, power: float):
-        point = np.array([velocity, power])
-        return {f: float(self.interpolators[f](point)) for f in self.features}
-
-
-def compute_mean_and_std(defects: np.ndarray):
-    if len(defects) == 0:
-        y = 0.0
-        yerr = np.nan
-    elif len(defects) == 1:
-        y = defects[0]
-        yerr = defects[0]
-    else:
-        y = np.mean(defects)
-        yerr = np.std(defects, ddof=1) / np.sqrt(len(defects))
-
-    return y, yerr
+        point = np.array([[velocity, power]])
+        return {f: float(self.interpolators[f](point)[0]) for f in self.features}
 
 
 def run_raptor(
@@ -135,26 +117,21 @@ def run_raptor(
     mp_interpolator: MeltPoolInterpolator,
     layer_thickness_m: float = 40e-6,
     query_volume_mm3: float = 1.0,
-    voxel_resolution_m: float = 5.0e-6,
+    voxel_resolution_m: float = 5e-6,
     metric_names: list[str] = ["equivalent_diameter_area"],
 ):
-    """Executes a single Raptor simulation using dynamic melt pool surrogates."""
-    # Re-seed at the start of every ensemble for independent determinism
     random.seed(SEED)
     np.random.seed(SEED)
     seed_numba(SEED)
 
-    # 1. Query the surrogates for power and velocity combination
     mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
 
-    # 2. Create voxel grid
     rve_min_point = np.array([0.0, 0.0, 0.0])
-    rve_max_point = np.array([5e-4, 5e-4, 5e-4])
+    rve_max_point = np.array([1e-3, 1e-3, 1e-3])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
 
     grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
 
-    # 3. Build the laser scan path
     scan_path_builder = ScanPathBuilder(
         rve_bounding_box,
         LASER_POWER_WATTS,
@@ -168,66 +145,38 @@ def run_raptor(
     scan_path_builder.generate_layers()
[…]

[…]
