### 2026-04-20 fb4f483
add manual GP-hyperparameters and use newer server deployment
