### 2026-04-21 b958c9f
add hyperparameters to defect model
- remove log1p output processing (not needed?)
- add VOXEL_RESOLUTION handling, to deal with the fact that pores
  below it can not be found
- use appropriate acquisition function for exploration
