+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -59,10 +59,12 @@ BOUNDS = [[70e-6, 170e-6]]
 UNIT_BOUNDS = [[0.0, 1.0]]
 NUM_DIMS = len(BOUNDS)
 
-INITIAL_DATA_SIZE = 20
+INITIAL_DATA_SIZE = 5
 MAX_ITERATIONS = 35
 
 SEED = 42
+
+VOXEL_RESOLUTION = 1e-5  # increase voxel_resolution to speed up
 
 
 @njit
@@ -116,7 +118,7 @@ def run_raptor(
     hatch_spacing_m: float,
     mp_interpolator: MeltPoolInterpolator,
     layer_thickness_m: float = 40e-6,
-    query_volume_mm3: float = 1.0,
+    query_volume_mm3: float = 1.0,  # is not used curently?
     voxel_resolution_m: float = 5e-6,
     metric_names: list[str] = ["equivalent_diameter_area"],
 ):
@@ -167,16 +169,38 @@ def run_raptor(
     porosity = compute_porosity(grid, path_vectors, melt_pool, False)
     metrics = compute_morphology(porosity, grid.resolution, metric_names)
 
-    combined_defects = metrics["equivalent_diameter_area"]
+    inputs = {
+        "hatch_spacing_m": hatch_spacing_m,
+        "layer_thickness_m": layer_thickness_m,
+        "query_volume_mm3": query_volume_mm3,
+        "voxel_resolution_m": voxel_resolution_m,
+    }
+
+    raptor_data = {"inputs": inputs,
+                   "outputs": metrics}
+
+    return raptor_data
+
+
+def process_raptor_data(raptor_data):
+
+    combined_defects = raptor_data["outputs"]["equivalent_diameter_area"]
+
+    # add 5 pores below the VOXEL_RESOLUTION, to deal with empty lists due to finite resolution
+    # TODO: discuss this, if we want to take a maximum here, we need to think about the extreme value statistics
+    combined_defects = combined_defects.tolist() + [VOXEL_RESOLUTION * 0.9] * 5
 
     if len(combined_defects) > 0:
         max_pore = np.max(combined_defects)
+        std_pore = np.std(combined_defects, ddof=1)
     else:
-        max_pore = 0.0
+         # if no pores are fouund, they must be below VOXEL_RESOLUTION
+        max_pore = VOXEL_RESOLUTION
+        std_pore = max_pore
 
-    logger.info(f"Hatch: {hatch_spacing_m*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um")
+    logger.info(f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um")
 
-    return float(np.log1p(max_pore))
+    return float(max_pore), float(std_pore)
 
 
 # -----------------------------------------------------------------------------
@@ -198,8 +222,11 @@ def x_from_unit(U):
 
 def get_data_point(x_suggested, mp_interpolator):
     x_ = [np.clip(x_suggested[0], BOUNDS[0][0], BOUNDS[0][1])]
-    y_ = run_raptor(x_[0], mp_interpolator)
-    return x_, y_
+    raptor_data = run_raptor(x_[0], mp_interpolator,
+                             voxel_resolution_m=VOXEL_RESOLUTION,
+                             )
+    y_ = process_raptor_data(raptor_data)[0]
+    return x_, y_, raptor_data
 
 
 # -----------------------------------------------------------------------------
@@ -221,8 +248,13 @@ class ActiveLearningOrchestrator:
         bounds_points = np.array([[BOUNDS[0][0]], [BOUNDS[0][1]]])
 
[…]
