### 2026-04-21 8f206c3
add heteroscedastic support
- process a y_err value
  (TODO: discuss extreme value statistics for max)
- pass y_err values through to the backend
