+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -55,17 +55,17 @@ logger = logging.getLogger(__name__)
 LASER_POWER_WATTS = 200.0
 LASER_VELOCITY_M_S = 1.0
 
-BOUNDS = [[70e-6, 170e-6]]
-UNIT_BOUNDS = [[0.0, 1.0]]
+BOUNDS = ((70e-6, 140e-6),) #((70e-6, 170e-6))
+UNIT_BOUNDS = ((0.0, 1.0),)
 NUM_DIMS = len(BOUNDS)
 
 INITIAL_DATA_SIZE = 5
-MAX_ITERATIONS = 35
+MAX_ITERATIONS = 30
 
 SEED = 42
 
-VOXEL_RESOLUTION = 1e-5  # increase voxel_resolution to speed up
-
+VOXEL_RESOLUTION_M = 2e-6  # increase voxel_resolution to speed up
+RVE_LENGTH_M       = 5e-4
 
 @njit
 def seed_numba(seed):
@@ -129,7 +129,7 @@ def run_raptor(
     mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
 
     rve_min_point = np.array([0.0, 0.0, 0.0])
-    rve_max_point = np.array([1e-3, 1e-3, 1e-3])
+    rve_max_point = np.array([RVE_LENGTH_M, RVE_LENGTH_M, RVE_LENGTH_M])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
 
     grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
@@ -184,18 +184,19 @@ def run_raptor(
 
 def process_raptor_data(raptor_data):
 
+    voxel_resolution_m = raptor_data["inputs"]["voxel_resolution_m"]
     combined_defects = raptor_data["outputs"]["equivalent_diameter_area"]
 
-    # add 5 pores below the VOXEL_RESOLUTION, to deal with empty lists due to finite resolution
+    # add 3 pores below the voxel_resolution, to deal with empty lists due to finite resolution
     # TODO: discuss this, if we want to take a maximum here, we need to think about the extreme value statistics
-    combined_defects = combined_defects.tolist() + [VOXEL_RESOLUTION * 0.9] * 5
+    combined_defects = combined_defects.tolist() + (voxel_resolution_m * np.random.rand(3)).tolist()
 
-    if len(combined_defects) > 0:
+    if len(combined_defects) > 1:
         max_pore = np.max(combined_defects)
-        std_pore = np.std(combined_defects, ddof=1)
+        std_pore = np.std(combined_defects, ddof=1) / np.sqrt(len(combined_defects))
     else:
          # if no pores are fouund, they must be below VOXEL_RESOLUTION
-        max_pore = VOXEL_RESOLUTION
+        max_pore = np.max(combined_defects, initial=voxel_resolution_m)
         std_pore = max_pore
 
     logger.info(f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um")
@@ -223,10 +224,10 @@ def x_from_unit(U):
 def get_data_point(x_suggested, mp_interpolator):
     x_ = [np.clip(x_suggested[0], BOUNDS[0][0], BOUNDS[0][1])]
     raptor_data = run_raptor(x_[0], mp_interpolator,
-                             voxel_resolution_m=VOXEL_RESOLUTION,
+                             voxel_resolution_m=VOXEL_RESOLUTION_M,
                              )
-    y_ = process_raptor_data(raptor_data)[0]
-    return x_, y_, raptor_data
+    y_, yerr_ = process_raptor_data(raptor_data)
[…]
