### 2026-04-24 b2b522e
add sable backend, nonlinear output normalization, and statistics
- add sable backend as alternative to sklearn
- add nonlinear output normalization to scale errors to log scale for
  acquisition function
- tweak and improve the way statistics are computed
