+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -64,8 +64,12 @@ MAX_ITERATIONS = 30
 
 SEED = 42
 
-VOXEL_RESOLUTION_M = 2e-6  # increase voxel_resolution to speed up
+VOXEL_RESOLUTION_M = 5e-6  # increase voxel_resolution to speed up
 RVE_LENGTH_M       = 5e-4
+
+MIN_LEN_DEFECTS    = 10
+ANALYZE_MAX        = False
+BACKEND            = "sklearn" #"sable"
 
 @njit
 def seed_numba(seed):
@@ -118,16 +122,22 @@ def run_raptor(
     hatch_spacing_m: float,
     mp_interpolator: MeltPoolInterpolator,
     layer_thickness_m: float = 40e-6,
-    query_volume_mm3: float = 1.0,  # is not used curently?
+    query_volume_mm3: float = 1.0,  # is not used currently?
     voxel_resolution_m: float = 5e-6,
     metric_names: list[str] = ["equivalent_diameter_area"],
 ):
-    random.seed(SEED)
-    np.random.seed(SEED)
-    seed_numba(SEED)
+    # query volume: should be set to the value from size of the experimental coupon
+    #    (artifact for process quality control)
+
+    # TODO: why do we do the seeding in every iteration?
+    # this is causing issues with the statistics, I commented it out.
+    #random.seed(SEED)
+    #np.random.seed(SEED)
+    #seed_numba(SEED)
 
     mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
 
+    # TODO: run enough RVSs to cover the whole query volume.
     rve_min_point = np.array([0.0, 0.0, 0.0])
     rve_max_point = np.array([RVE_LENGTH_M, RVE_LENGTH_M, RVE_LENGTH_M])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
@@ -187,47 +197,82 @@ def process_raptor_data(raptor_data):
     voxel_resolution_m = raptor_data["inputs"]["voxel_resolution_m"]
     combined_defects = raptor_data["outputs"]["equivalent_diameter_area"]
 
-    # add 3 pores below the voxel_resolution, to deal with empty lists due to finite resolution
-    # TODO: discuss this, if we want to take a maximum here, we need to think about the extreme value statistics
-    combined_defects = combined_defects.tolist() + (voxel_resolution_m * np.random.rand(3)).tolist()
+    min_len_defects = MIN_LEN_DEFECTS
+    if len(combined_defects) < min_len_defects:
+        # add 3 pores below the voxel_resolution, to deal with empty lists due to finite resolution
+        # TODO: discuss how we should handle the fact that pores blow voxel_resolution can not be resolved
+        random.seed() # explicitly call rng seeding to make sure this is truly random
+        n_extra_defects = min_len_defects - len(combined_defects)
+        more_defects = (voxel_resolution_m * np.random.rand(n_extra_defects)).tolist()
+        combined_defects = combined_defects.tolist() + more_defects
 
-    if len(combined_defects) > 1:
-        max_pore = np.max(combined_defects)
-        std_pore = np.std(combined_defects, ddof=1) / np.sqrt(len(combined_defects))
+    max_pore = np.max(combined_defects)
+    mean_pore = np.mean(combined_defects)
+    std_pore = np.std(combined_defects, ddof=1)
+    # compute the standard error, the standard deviation of the mean (Monte-Carlo error)
+    std_err_pore = std_pore / np.sqrt(len(combined_defects))
+
+    logger.info(f"Found {len(combined_defects)} defects: "
[…]
