### 2026-04-24 abee1e0
require newer version of dial in pyproject.toml
