+++ pyproject.toml
@@ -30,7 +30,7 @@ dependencies = [
     "pandas",
     "pyvista",
     "intersect-sdk",
-    "dial @ git+https://github.com/INTERSECT-DIAL/dial"
+    "dial @ git+https://github.com/INTERSECT-DIAL/dial@initial_benchmarking"
 ]
 
 [project.scripts]
