+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -55,7 +55,7 @@ logger = logging.getLogger(__name__)
 LASER_POWER_WATTS = 200.0
 LASER_VELOCITY_M_S = 1.0
 
-BOUNDS = ((70e-6, 140e-6),) #((70e-6, 170e-6))
+BOUNDS = ((70e-6, 140e-6),)  # ((70e-6, 170e-6))
 UNIT_BOUNDS = ((0.0, 1.0),)
 NUM_DIMS = len(BOUNDS)
 
@@ -65,11 +65,12 @@ MAX_ITERATIONS = 30
 SEED = 42
 
 VOXEL_RESOLUTION_M = 5e-6  # increase voxel_resolution to speed up
-RVE_LENGTH_M       = 5e-4
+RVE_LENGTH_M = 5e-4
 
-MIN_LEN_DEFECTS    = 10
-ANALYZE_MAX        = False
-BACKEND            = "sklearn" #"sable"
+MIN_LEN_DEFECTS = 10
+ANALYZE_MAX = False
+BACKEND = "sklearn"  # "sable"
+
 
 @njit
 def seed_numba(seed):
@@ -131,9 +132,9 @@ def run_raptor(
 
     # TODO: why do we do the seeding in every iteration?
     # this is causing issues with the statistics, I commented it out.
-    #random.seed(SEED)
-    #np.random.seed(SEED)
-    #seed_numba(SEED)
+    # random.seed(SEED)
+    # np.random.seed(SEED)
+    # seed_numba(SEED)
 
     mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
 
@@ -186,8 +187,7 @@ def run_raptor(
         "voxel_resolution_m": voxel_resolution_m,
     }
 
-    raptor_data = {"inputs": inputs,
-                   "outputs": metrics}
+    raptor_data = {"inputs": inputs, "outputs": metrics}
 
     return raptor_data
 
@@ -201,7 +201,7 @@ def process_raptor_data(raptor_data):
     if len(combined_defects) < min_len_defects:
         # add 3 pores below the voxel_resolution, to deal with empty lists due to finite resolution
         # TODO: discuss how we should handle the fact that pores blow voxel_resolution can not be resolved
-        random.seed() # explicitly call rng seeding to make sure this is truly random
+        random.seed()  # explicitly call rng seeding to make sure this is truly random
         n_extra_defects = min_len_defects - len(combined_defects)
         more_defects = (voxel_resolution_m * np.random.rand(n_extra_defects)).tolist()
         combined_defects = combined_defects.tolist() + more_defects
@@ -212,10 +212,12 @@ def process_raptor_data(raptor_data):
     # compute the standard error, the standard deviation of the mean (Monte-Carlo error)
     std_err_pore = std_pore / np.sqrt(len(combined_defects))
 
-    logger.info(f"Found {len(combined_defects)} defects: "
-                f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um")
+    logger.info(
+        f"Found {len(combined_defects)} defects: "
+        f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um"
+    )
 
-    # TODO: discuss and refine the data analyis and extreme value statistics
+    # TODO: discuss and refine the data analysis and extreme value statistics
     return_mean = not ANALYZE_MAX
     if return_mean:
         y, yerr = float(mean_pore), float(std_err_pore)
@@ -245,15 +247,16 @@ def x_from_unit(x):
 
[…]
