### 2026-04-24 576a37e
Improve dial surrogate (#30)
* add manual GP-hyperparameters and use newer server deployment

* add hyperparameters to defect model

- remove log1p output processing (not needed?)
- add VOXEL_RESOLUTION handling, to deal with the fact that pores
  below it can not be found
- use appropriate acquisition function for exploration

* add heteroscedastic support

- process a y_err value
  (TODO: discuss extreme value statistics for max)
- pass y_err values through to the backend

* add sable backend, nonlinear output normalization, and statistics

- add sable backend as alternative to sklearn
- add
[…]
