+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -55,14 +55,21 @@ logger = logging.getLogger(__name__)
 LASER_POWER_WATTS = 200.0
 LASER_VELOCITY_M_S = 1.0
 
-BOUNDS = [[70e-6, 170e-6]]
-UNIT_BOUNDS = [[0.0, 1.0]]
+BOUNDS = ((70e-6, 140e-6),)  # ((70e-6, 170e-6))
+UNIT_BOUNDS = ((0.0, 1.0),)
 NUM_DIMS = len(BOUNDS)
 
-INITIAL_DATA_SIZE = 20
-MAX_ITERATIONS = 35
+INITIAL_DATA_SIZE = 5
+MAX_ITERATIONS = 30
 
 SEED = 42
+
+VOXEL_RESOLUTION_M = 5e-6  # increase voxel_resolution to speed up
+RVE_LENGTH_M = 5e-4
+
+MIN_LEN_DEFECTS = 10
+ANALYZE_MAX = False
+BACKEND = "sklearn"  # "sable"
 
 
 @njit
@@ -116,18 +123,24 @@ def run_raptor(
     hatch_spacing_m: float,
     mp_interpolator: MeltPoolInterpolator,
     layer_thickness_m: float = 40e-6,
-    query_volume_mm3: float = 1.0,
+    query_volume_mm3: float = 1.0,  # is not used currently?
     voxel_resolution_m: float = 5e-6,
     metric_names: list[str] = ["equivalent_diameter_area"],
 ):
-    random.seed(SEED)
-    np.random.seed(SEED)
-    seed_numba(SEED)
+    # query volume: should be set to the value from size of the experimental coupon
+    #    (artifact for process quality control)
+
+    # TODO: why do we do the seeding in every iteration?
+    # this is causing issues with the statistics, I commented it out.
+    # random.seed(SEED)
+    # np.random.seed(SEED)
+    # seed_numba(SEED)
 
     mp_stats = mp_interpolator.query(LASER_VELOCITY_M_S, LASER_POWER_WATTS)
 
+    # TODO: run enough RVSs to cover the whole query volume.
     rve_min_point = np.array([0.0, 0.0, 0.0])
-    rve_max_point = np.array([1e-3, 1e-3, 1e-3])
+    rve_max_point = np.array([RVE_LENGTH_M, RVE_LENGTH_M, RVE_LENGTH_M])
     rve_bounding_box = np.array([rve_min_point, rve_max_point])
 
     grid = create_grid(voxel_resolution=voxel_resolution_m, bound_box=rve_bounding_box)
@@ -167,39 +180,101 @@ def run_raptor(
     porosity = compute_porosity(grid, path_vectors, melt_pool, False)
     metrics = compute_morphology(porosity, grid.resolution, metric_names)
 
-    combined_defects = metrics["equivalent_diameter_area"]
+    inputs = {
+        "hatch_spacing_m": hatch_spacing_m,
+        "layer_thickness_m": layer_thickness_m,
+        "query_volume_mm3": query_volume_mm3,
+        "voxel_resolution_m": voxel_resolution_m,
+    }
 
-    if len(combined_defects) > 0:
-        max_pore = np.max(combined_defects)
+    raptor_data = {"inputs": inputs, "outputs": metrics}
+
+    return raptor_data
+
+
+def process_raptor_data(raptor_data):
+
+    voxel_resolution_m = raptor_data["inputs"]["voxel_resolution_m"]
+    combined_defects = raptor_data["outputs"]["equivalent_diameter_area"]
+
+    min_len_defects = MIN_LEN_DEFECTS
+    if len(combined_defects) < min_len_defects:
+        # add 3 pores below the voxel_resolution, to deal with empty lists due to finite resolution
[…]
