### 2026-05-01 dc3c603
added ensemble rve sampling
