### 2026-05-27 893c451
melt pool boundary detection implementation
