### 2026-05-28 69038b1
simplified boundary detection. need to add an option for tolerance on boundary.
