### 2026-05-29 37cf6f2
implemented newton iterations for boundary detection. need to add boundary width as a user-controlled parameter.
