+++ src/raptor/core.py
@@ -106,6 +106,61 @@ def is_boundary(
 
     return np.abs(test_value - 1.0) <= tolerance
 
+@njit(fastmath=True)
+def compute_distance_to_boundary(
+    y: float,
+    z: float,
+    width: float,
+    height: float,
+    depth: float,
+    height_shape_factor: float,
+    depth_shape_factor: float,
+) -> float:
+    """
+    Computes the distance from a point (y, z) to the boundary of a modified Lamé curve cross-section:
+    (y/a)^2 + (|z|/b)^n <= 1
+
+    Args:
+        y (float): The y-coordinate of the point (horizontal axis).
+        z (float): The z-coordinate of the point (vertical axis).
+        width (float): The full width of the shape along the y-axis (shared).
+        height (float): The maximum height of the top half of the shape (for z > 0).
+        depth (float): The maximum depth of the bottom half of the shape (for z < 0).
+        height_shape_factor (float): The shape exponent 'n' for the top half.
+        depth_shape_factor (float): The shape exponent 'n' for the bottom half.
+            - n=2:   Ellipse
+            - n=1:   Parabola
+            - n=0.5: Bell-shaped
+            - n=10:  Box-shaped
+
+    Returns:
+        float: The distance from the point to the boundary. Positive if outside, negative if inside.
+    """
+
+    a = width / 2.0
+
+    b_choices = (depth, height)
+    n_choices = (depth_shape_factor, height_shape_factor)
+
+    selector = int(z >= 0)
+
+    b = b_choices[selector]
+    n = n_choices[selector]
+
+    theta = np.arctan2(np.abs(z), y)
+    cos_theta = np.cos(theta)
+    sin_theta = np.sin(theta)
+    r0 = (y**2 + z**2) ** 0.5
+    rtol = 1e-8
+    while True:
+        f = (r0 * cos_theta / a) ** 2 + (r0 * sin_theta / b) ** n - 1.0
+        if np.abs(f) < rtol:
+            break
+        df_dr = 2 * (r0 * cos_theta / a) ** 2 / r0 + n * (r0 * sin_theta / b) ** n / r0
+        r0 -= f / df_dr
+
+    return r0
+
 def compute_melt_mask(
     voxels: np.ndarray, melt_pool: MeltPool, path_vectors: List[PathVector]
 ):
@@ -292,7 +347,10 @@ def compute_melt_mask_implicit(
             
             is_voxel_melted = is_inside(local_y, local_z, width, height, depth, height_shape_factor, depth_shape_factor)
             
-            is_voxel_boundary = is_boundary(local_y, local_z, width, height, depth, height_shape_factor, depth_shape_factor, 0.1)
+            # is_voxel_boundary = is_boundary(local_y, local_z, width, height, depth, height_shape_factor, depth_shape_factor, 0.1)
+
+            dist_to_bdry = compute_distance_to_boundary(local_y, local_z, width, height, depth, height_shape_factor, depth_shape_factor)
+            is_voxel_boundary = np.abs(dist_to_bdry - (local_y**2 + local_z**2)**0.5) <= 5.0e-6 * (2**0.5)/2
 
             if is_voxel_melted:
                  melt_mask[i] = 1
