### 2026-05-29 833db70
added resolution argument to compute_melt_mask, etc. currently using grid.resolution for this. need to find a clean way to specify boundary width.
