### 2026-06-01 73c3bc0
newton iteration implementation.
