+++ src/raptor/api.py
@@ -176,7 +176,7 @@ def compute_porosity(
 
         # Warm up the main, parallelized compute kernel.
         if grid.n_voxels > 0 and path_vectors:
-            _ = compute_melt_mask(grid.voxels[0:1], grid.resolution, melt_pool, path_vectors[0:1])
+            _ = compute_melt_mask(grid.voxels[0:1], grid.resolution * 1.5, melt_pool, path_vectors[0:1])
 
         print(f" -> JIT warmup complete ({time.time() - t_start_warmup:.8f}s).")
 
@@ -188,7 +188,7 @@ def compute_porosity(
 
     print("Running melt-mask calculation...")
     t0_run = time.time()
-    melted_mask_flat = compute_melt_mask(grid.voxels, grid.resolution, melt_pool, path_vectors)
+    melted_mask_flat = compute_melt_mask(grid.voxels, grid.resolution * 1.5, melt_pool, path_vectors)
     t_elapsed = time.time() - t0_run
 
     n_melted = melted_mask_flat.sum()
