### 2026-06-05 e00b21e
cleaned up newton solver
