### 2026-06-05 54d7620
removed old is_boundary
