### 2026-06-11 e96180b
Update dial dependency to use 'develop' branch
