### 2026-06-11 369bce2
intersect local config example using docker compose
