### 2026-06-17 5bdc65c
added boundary tolerance parameter for newton's method convergence
