### 2026-06-17 b1c9cf5
cleaned up implementation to only use resolution and some geometry
