+++ examples/api_example/rve.py
@@ -27,7 +27,7 @@ from raptor.utilities import ScanPathBuilder
 min_point = np.array([0.0, 0.0, 0.0])
 max_point = np.array([5.0e-4, 5.0e-4, 5.0e-4])
 bound_box = np.array([min_point, max_point])
-voxel_resolution = 5.0e-6
+voxel_resolution = 2.5e-6
 
 grid = create_grid(voxel_resolution, bound_box=bound_box)
 
@@ -81,8 +81,7 @@ melt_pool_dict = {
 melt_pool = create_melt_pool(melt_pool_dict, enable_random_phases=True)
 
 # 4. Compute porosity using conic section / superellipse curves for melt pool mask
-boundary_tolerance = 1e-5 # setting a convergence tolerance for the boundary calculation in compute_melt_mask()
-porosity = compute_porosity(grid, path_vectors, melt_pool, jit_warmup=True, boundary_tolerance=boundary_tolerance)
+porosity = compute_porosity(grid, path_vectors, melt_pool, jit_warmup=True)
 
 # 5. Write porosity field to .VTI
 write_vtk(grid.origin, grid.resolution, porosity, "rve.vti")
