### 2026-06-18 b3a69b3
some minor params i changed, changing back. testing still fine
