### 2026-06-18 6f6f0c0
typo in newton. redid it, same results.
