### 2026-06-18 76f0515
faster newton, simplified voxel checks by using signed distance field.
