### 2026-06-18 12611ef
updated readme with animation.
