### 2026-06-18 5244b0c
modified compute_morphology to handle new data format.
