### 2026-06-18 bf5c6e7
reset to old skimage min_size argument for morphology. expect deprecation warning. function still works as intended
