+++ src/raptor/core.py
@@ -275,16 +275,6 @@ def compute_melt_mask_implicit(
                     two_pi_t * height_frequencies[k] + phase_k
                 )
 
-            # is_voxel_melted = is_inside(
-            #     local_y,
-            #     local_z,
-            #     width,
-            #     height,
-            #     depth,
-            #     height_shape_factor,
-            #     depth_shape_factor,
-            # )
-
             signed_dist = compute_distance_to_boundary(
                 local_y,
                 local_z,
@@ -308,4 +298,5 @@ def compute_melt_mask_implicit(
                 melt_mask[i] = 2
             if melt_mask_previous > 1 and is_voxel_boundary:
                 melt_mask[i] = 3
+            
     return melt_mask
