### 2026-06-18 8a67bb8
fixed plotting. changed rve.py example to create minimal visualizable porosity reliably with increased hatch and layer thickness.
