### 2026-06-18 a8f5dac
fixed visualization error flag. exits after printing defect status.
