+++ src/raptor/api.py
@@ -293,7 +293,12 @@ def visualize(vtk_output_path: str, scaling=1e6) -> None:
     Defaults to scaling from meters to microns for better labeling.
     """
     rve = pv.read(vtk_output_path)
-    isosurface = rve.contour(isosurfaces=[0.5], scalars="porosity", compute_normals=False)
+    isosurface = rve.contour(
+        isosurfaces=[0.5], scalars="porosity", compute_normals=False
+    )
+    if isosurface.n_points == 0:
+        print(f"No porosity detected in the volume, skipping visualization.")
+        return
 
     # Outline of the original domain
     outline = rve.outline()
