+++ examples/mpi_example/test_mpi.py
@@ -1,0 +1,167 @@
+# test_mpi.py
+import numpy as np
+from mpi4py import MPI
+import h5py
+from raptor.api import (
+    create_path_vectors,
+    create_melt_pool,
+    create_grid,
+    compute_porosity,
+    write_vtk,
+    compute_morphology,
+    write_morphology,
+    visualize,
+)
+from raptor.utilities import ScanPathBuilder, MeltPoolFilter
+
+comm = MPI.COMM_WORLD
+rank, size = comm.Get_rank(), comm.Get_size()
+
+
+def run_raptor(width, depth, sigma=0, edgelength=5.0e-4):
+
+    
+    # 1. Create voxel grid for the representative volume element (RVE)
+    min_point = np.array([0.0, 0.0, 0.0])
+    edgelength = np.atleast_1d(edgelength)
+    if len(edgelength) == 1:
+        max_point = np.array([edgelength[0], edgelength[0], edgelength[0]])
+    elif len(edgelength) == 3:
+        max_point = np.array([edgelength[0], edgelength[1], edgelength[2]])
+    else:
+        raise ValueError("edgelength must be a scalar or a list/array of three values.")
+    bound_box = np.array([min_point, max_point])
+    voxel_resolution = 5.0e-6
+
+    grid = create_grid(voxel_resolution, bound_box=bound_box)
+
+    # 2. Create path vectors through the representative volume element (RVE)
+    power = 370
+    velocity = 1.7
+    hatch_spacing = 140e-6
+    layer_height = 30e-6
+    rotation = 67
+    scan_extension = max(max_point - min_point)
+    extra_layers = 10
+
+    scan_path_builder = ScanPathBuilder(
+        bound_box,
+        power,
+        velocity,
+        hatch_spacing,
+        layer_height,
+        rotation,
+        scan_extension,
+        extra_layers,
+    )
+
+    scan_path_builder.generate_layers()
+    path_vectors = scan_path_builder.process_vectors()
+
+    # 3. Create melt pools from convolution filter
+    mean = width
+    std_dev = sigma if sigma > 0 else 0.01e-6
+    frequency = 250000
+
+    duration = 0.08
+
+    # Instantiate object
+    mp_w_filter = MeltPoolFilter(width, std_dev, velocity, [frequency, duration])
+    mp_d_filter = MeltPoolFilter(depth, std_dev, velocity, [frequency, duration])
+
+    # Define physical scales
+    mp_w_filter.add_effect("melt_pool", [500e-6, None, 1])
+    mp_d_filter.add_effect("melt_pool", [500e-6, None, 1])
+
+    # Generate stochastic melt pool
+    mp_w_filter.initialize()
+    width_data = mp_w_filter.generate_fluctuations(1)
+
+    mp_d_filter.initialize()
+    depth_data = mp_d_filter.generate_fluctuations(1)
+    
+    n_modes = 50
+
+    # scale melt pool data by constant factor
+    width_scale = 1.0
+    depth_scale = 1.0
+    height_scale = 0.4
+
+    # assign shape to melt pool and cap (1 = parabola, 2 = ellipse)
+    width_shape = 2  # placeholder
+    height_shape = 1
+    depth_shape = 1
+
+    melt_pool_dict = {
+        "width": (width_data, n_modes, width_scale, width_shape),
+        "depth": (depth_data, n_modes, depth_scale, depth_shape),
+        "height": (width_data, n_modes, height_scale, height_shape),
+    }
+
+    melt_pool = create_melt_pool(melt_pool_dict, enable_random_phases=True)
+
+    # 4. Compute porosity using conic section / superellipse curves for melt pool mask
[…]
