### 2026-06-18 49b2c88
readme updated with gif.
