+++ examples/mpi_example/test_mpi.py
@@ -13,13 +13,41 @@ from raptor.api import (
     visualize,
 )
 from raptor.utilities import ScanPathBuilder, MeltPoolFilter
+import joblib
+import pickle
 
 comm = MPI.COMM_WORLD
 rank, size = comm.Get_rank(), comm.Get_size()
+INITIAL_BOUNDS = ((0.5, 1.5), (100, 300))  # [velocity (m/s), power (W)]
+UNIT_BOUNDS = ((0.0, 1.0), (0.0, 1.0))
+
+# scaling helper functions
+def x_to_unit(X):
+    X = np.asarray(X, dtype=float)
+    lo = np.array([b[0] for b in INITIAL_BOUNDS])
+    hi = np.array([b[1] for b in INITIAL_BOUNDS])
+    return (X - lo) / (hi - lo + 1e-12)
 
 
-def run_raptor(width, depth, sigma=0, edgelength=5.0e-4):
+def x_from_unit(U):
+    U = np.asarray(U, dtype=float)
+    lo = np.array([b[0] for b in INITIAL_BOUNDS])
+    hi = np.array([b[1] for b in INITIAL_BOUNDS])
+    return U * (hi - lo) + lo
 
+
+def load_meltpool_surrogates(joblib_path):
+    # Loading pretrained gaussian process surrogates for melt pool dimensions and variability.
+    meltpool_surrogates = joblib.load(joblib_path)
+    return meltpool_surrogates
+
+def load_meltpool_data(pickle_path):
+    with open(pickle_path, "rb") as f:
+        meltpool_data = pickle.load(f)
+    return meltpool_data
+
+def run_raptor(power, velocity, hatch_spacing, layer_height, features, 
+               edgelength=5.0e-4):    
     
     # 1. Create voxel grid for the representative volume element (RVE)
     min_point = np.array([0.0, 0.0, 0.0])
@@ -36,10 +64,10 @@ def run_raptor(width, depth, sigma=0, edgelength=5.0e-4):
     grid = create_grid(voxel_resolution, bound_box=bound_box)
 
     # 2. Create path vectors through the representative volume element (RVE)
-    power = 370
-    velocity = 1.7
-    hatch_spacing = 140e-6
-    layer_height = 30e-6
+    power = power
+    velocity = velocity
+    hatch_spacing = hatch_spacing
+    layer_height = layer_height
     rotation = 67
     scan_extension = max(max_point - min_point)
     extra_layers = 10
@@ -59,19 +87,26 @@ def run_raptor(width, depth, sigma=0, edgelength=5.0e-4):
     path_vectors = scan_path_builder.process_vectors()
 
     # 3. Create melt pools from convolution filter
-    mean = width
-    std_dev = sigma if sigma > 0 else 0.01e-6
+    mu_w = features["width_mean"]
+    sig_w = features["width_std"]
+    mu_d = features["depth_mean"]
+    sig_d = features["depth_std"]
+    mu_h = features["height_mean"]
+    sig_h = features["height_std"]
+
     frequency = 250000
 
     duration = 0.08
 
     # Instantiate object
-    mp_w_filter = MeltPoolFilter(width, std_dev, velocity, [frequency, duration])
-    mp_d_filter = MeltPoolFilter(depth, std_dev, velocity, [frequency, duration])
+    mp_w_filter = MeltPoolFilter(mu_w, sig_w, velocity, [frequency, duration])
+    mp_d_filter = MeltPoolFilter(mu_d, sig_d, velocity, [frequency, duration])
+    mp_h_filter = MeltPoolFilter(mu_h, sig_h, velocity, [frequency, duration])
 
     # Define physical scales
[…]
