### 2026-06-19 2727d75
fixed some visualization issues
