### 2026-06-24 190d10f
fixed issue by enforcing lowerbound on vector components at construction time.
