+++ src/raptor/core.py
@@ -286,7 +286,7 @@ def compute_melt_mask_implicit(
                 resolution,
             )
 
-            is_voxel_melted = signed_dist < 0.0
+            is_voxel_melted = signed_dist < 1e-24
 
             is_voxel_boundary = np.abs(signed_dist) - resolution <= 1e-24
 
