+++ src/raptor/structures.py
@@ -113,6 +113,7 @@ class PathVector:
 
     def set_coordinate_frame(self) -> None:
         self.distance = self.end_point - self.start_point
+        self.distance = np.where(np.abs(self.distance) < 1e-12, 1e-12, self.distance)
         self.centroid = (self.end_point + self.start_point) / 2.0
 
         self.duration = self.end_time - self.start_time
