### 2026-06-24 e261976
better implementation of signed clip + fix tol in core
