### 2026-06-30 16a8db7
fixed 0-indexing issue in self.t size and self.npoints discrepancy
