### 2026-06-30 0251ce8
changed to linspace to enforce same number of points. checked that dt was correct
