### 2026-07-01 c72def6
removed transcendental cosine computation
