+++ src/raptor/core.py
@@ -56,9 +56,6 @@ def compute_distance_to_boundary(
     b = b_choices[selector]
     n = n_choices[selector]
 
-    theta = np.arctan2(np.abs(z), y)
-    cos_theta = np.cos(theta)
-    sin_theta = np.sin(theta)
     rv = (y**2 + z**2) ** 0.5
     r0 = rv
     inv_r0 = 1.0 / r0
