### 2026-07-16 90d3e8a
Fixup for time scale in synthetic melt pool
