+++ src/raptor/api.py
@@ -68,28 +68,54 @@ def create_path_vectors(
 def compute_spectral_components(melt_pool_data: np.ndarray, n_modes: int) -> np.ndarray:
 
     dt = melt_pool_data[1, 0] - melt_pool_data[0, 0]
-    mode0 = melt_pool_data[:, 1].mean()
-    fft_resolution = np.fft.fft(melt_pool_data[:, 1])
-    F = np.zeros_like(fft_resolution)
-    n_fft = len(fft_resolution)
+    values = melt_pool_data[:, 1]
+    mode0 = values.mean()
+    n_fft = len(values)
     if n_modes == 1:
         spectral_array = np.array([[mode0, 0, 0]])
     else:
-        for i in range(1, n_modes):
-            F[i] = fft_resolution[i]
-            F[n_fft - i] = fft_resolution[n_fft - i]
+        fft_resolution = np.fft.rfft(values - mode0)
+        frequencies = np.fft.rfftfreq(n_fft, dt)
+        available_indices = np.arange(1, len(fft_resolution))
+        n_fluctuation_modes = min(n_modes - 1, len(available_indices))
 
-        frequencies = np.float64(1 / (dt * n_fft)) * np.arange(
-            n_modes, dtype=np.float64
-        )
-        phases = np.float64(np.angle(F[:n_modes]))
-        amplitudes = np.float64(np.abs(F[:n_modes]) / n_fft)
+        dominant = available_indices[
+            np.argsort(np.abs(fft_resolution[available_indices]))[
+                -n_fluctuation_modes:
+            ]
+        ]
+        dominant.sort()
+
+        selected_fft = fft_resolution[dominant]
+        selected_frequencies = frequencies[dominant]
+        amplitudes = 2.0 * np.abs(selected_fft) / n_fft
+
+        # The Nyquist term has no negative-frequency partner.
+        if n_fft % 2 == 0:
+            amplitudes[dominant == n_fft // 2] *= 0.5
+
+        # Preserve the variance of the input after truncating the spectrum.
+        represented_variance = 0.5 * np.sum(amplitudes**2)
+        if n_fft % 2 == 0:
+            represented_variance += 0.5 * np.sum(
+                amplitudes[dominant == n_fft // 2] ** 2
+            )
+        target_std = np.std(values)
+        if represented_variance > 0.0:
+            amplitudes *= target_std / np.sqrt(represented_variance)
+
+        phases = np.angle(selected_fft)
         spectral_array = np.vstack(
             [
                 np.array([mode0, 0, 0]),
-                np.vstack([amplitudes[1:], frequencies[1:], phases[1:]]).transpose(),
+                np.vstack([amplitudes, selected_frequencies, phases]).transpose(),
             ]
         )
+
+        if spectral_array.shape[0] < n_modes:
+            spectral_array = np.vstack(
+                [spectral_array, np.zeros((n_modes - spectral_array.shape[0], 3))]
+            )
     return np.float64(spectral_array)
 
 
