+++ tests/test_api.py
@@ -332,6 +332,35 @@ class TestComputeSpectralComponents:
 
         np.testing.assert_allclose(spectral_array[0, 0], expected_mean)
 
+    def test_compute_spectral_components_preserves_standard_deviation(
+        self, sample_time_series_data
+    ):
+        """Test that the truncated cosine representation preserves variance."""
+        spectral_array = compute_spectral_components(sample_time_series_data, 3)
+        time = sample_time_series_data[:, 0]
+        reconstructed = np.zeros_like(time)
+
+        for amplitude, frequency, phase in spectral_array:
+            reconstructed += amplitude * np.cos(2.0 * np.pi * frequency * time + phase)
+
+        assert reconstructed.mean() == pytest.approx(
+            sample_time_series_data[:, 1].mean()
+        )
+        assert reconstructed.std() == pytest.approx(
+            sample_time_series_data[:, 1].std()
+        )
+
+    def test_compute_spectral_components_selects_dominant_modes(self):
+        """Test that high-frequency signal content is not discarded."""
+        sampling_frequency = 10_000.0
+        time = np.arange(1000) / sampling_frequency
+        values = 1.0 + 0.2 * np.sin(2.0 * np.pi * 2000.0 * time)
+        spectral_array = compute_spectral_components(
+            np.column_stack([time, values]), 2
+        )
+
+        assert spectral_array[1, 1] == pytest.approx(2000.0)
+
     def test_compute_spectral_components_invalid_input(self):
         """Test spectral component computation with invalid input."""
         with pytest.raises(IndexError):
