### 2026-07-16 562e069
Fixup for time scale in synthetic melt pool
