+++ tests/test_api.py
@@ -346,18 +346,14 @@ class TestComputeSpectralComponents:
         assert reconstructed.mean() == pytest.approx(
             sample_time_series_data[:, 1].mean()
         )
-        assert reconstructed.std() == pytest.approx(
-            sample_time_series_data[:, 1].std()
-        )
+        assert reconstructed.std() == pytest.approx(sample_time_series_data[:, 1].std())
 
     def test_compute_spectral_components_selects_dominant_modes(self):
         """Test that high-frequency signal content is not discarded."""
         sampling_frequency = 10_000.0
         time = np.arange(1000) / sampling_frequency
         values = 1.0 + 0.2 * np.sin(2.0 * np.pi * 2000.0 * time)
-        spectral_array = compute_spectral_components(
-            np.column_stack([time, values]), 2
-        )
+        spectral_array = compute_spectral_components(np.column_stack([time, values]), 2)
 
         assert spectral_array[1, 1] == pytest.approx(2000.0)
 
