### 2026-07-17 bd0f4f5
mpfilter fix for proper timeseries generation
