### 2026-07-17 1d45e52
fixed spectral component determination with rmse construction
