### 2026-07-17 a1be731
implemented lightweight mcmc for lognormal distribution estimation from the defect data out of raptor.
