+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -10,6 +10,7 @@ from typing import Any
 import numpy as np
 from scipy.stats import qmc
 from scipy.interpolate import RegularGridInterpolator
+import scipy.stats as st
 
 # Raptor Imports
 from numba import njit
@@ -233,6 +234,14 @@ def process_raptor_data(raptor_data):
         more_defects = (voxel_resolution_m * np.random.rand(n_extra_defects)).tolist()
         combined_defects = combined_defects.tolist() + more_defects
 
+    # estimate distribution parameters for pore size distribution
+    # using a lightweight mcmc approach assuming a lognormal underlying distribution
+    # target y - E[µ] in posterior, yerr - sqrt(Var[µ]) in posterior
+    # Converting to microns for numerical stability in MCMC
+    trace = run_metropolis_hastings(np.array(combined_defects)*1e6, iterations=5000, proposal_widths=np.array([1,1]))
+    burnin = 1000
+    trace = trace.T[ :, burnin:]  # discard burn-in samples
+
     max_pore = np.max(combined_defects)
     mean_pore = np.mean(combined_defects)
     std_pore = np.std(combined_defects, ddof=1)
@@ -242,12 +251,14 @@ def process_raptor_data(raptor_data):
     logger.info(
         f"Found {len(combined_defects)} defects: "
         f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um"
+        f"Estimated lognormal E[µ]: {np.mean(trace[0]):.6f} | sqrt(Var[µ]): {np.std(trace[0]):.6f}"
     )
 
     # TODO: discuss and refine the data analysis and extreme value statistics
     return_mean = not ANALYZE_MAX
     if return_mean:
-        y, yerr = float(mean_pore), float(std_err_pore)
+        # y, yerr = float(mean_pore), float(std_err_pore)
+        y, yerr = float(np.mean(trace[0])), float(np.std(trace[0]))
     else:
         # return the maximum pore size, use the standard deveiation as approximate error estimate
         y, yerr = float(max_pore), float(std_pore)
@@ -302,6 +313,47 @@ def y_from_unit(y_norm, yerr_norm, y_scale=(1.0, 1.0)):
     y = y_prescale * np.expm1(y_postscale * y_norm)
     yerr = yerr_norm * (y + y_prescale) * y_postscale
     return y.tolist(), yerr.tolist()
+
+def log_prior_lognormal(params):
+    mu, sigma = params
+    if sigma <= 0:
+        return -np.inf  # log(0)
+    mu_prior = st.norm.logpdf(mu, loc=0, scale=10)  # Example prior for mean
+    sigma_prior = st.norm.logpdf(sigma, loc=1, scale=5)  # Example prior for std
+    return mu_prior + sigma_prior
+
+def loglikelihood_lognormal(params, data):
+    mu, sigma = params
+    if sigma <= 0:
+        return -np.inf  # log(0)
+    return np.sum(st.lognorm.logpdf(data, s=sigma, scale=np.exp(mu)))
+
+def log_posterior_lognormal(params, data):
+    return loglikelihood_lognormal(params, data) + log_prior_lognormal(params)
+
+def run_metropolis_hastings(data, iterations=10000, proposal_widths=np.array([1.0, 2.0])):
+    # Initial guesses
+    current_params = np.array([1, 1])  # Example initial guess
+    current_log_post = log_posterior_lognormal(current_params, data)
+    
+    trace = []
+    
+    for i in range(iterations):
+        # Propose new parameters (Random Walk) 
+        proposal = current_params + np.random.normal(0, proposal_widths, size=current_params.shape)
+        
+        proposal_log_post = log_posterior_lognormal(proposal, data)
+        
[…]
