### 2026-07-17 4c374fd
use newer dial features to streamline (y,yerr) handling
- send (y,yerr) to the database (initial and update)
- adapt to newer dial development version @develop (get_surrogate_values)
- use new prior_std for the sable backend, requires the @sable branch
  for dial
