+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -39,6 +39,7 @@ from dial_dataclass import (
     DialInputSingleOtherStrategy,
     DialWorkflowCreationParamsClient,
     DialWorkflowDatasetUpdate,
+    Normal,
 )
 
 
@@ -58,18 +59,18 @@ BOUNDS = ((70e-6, 140e-6),)
 UNIT_BOUNDS = ((0.0, 1.0),)
 NUM_DIMS = len(BOUNDS)
 
-INITIAL_DATA_SIZE = 5
-MAX_ITERATIONS = 30
+INITIAL_DATA_SIZE = 1
+MAX_ITERATIONS = 20
 
 SEED = 42
 
 VOXEL_RESOLUTION_M = 5e-6
 RVE_LENGTH_M = 1e-3
-QUERY_VOLUME_MM3 = 10  # decrease query_volume_mm3 to speed up
+QUERY_VOLUME_MM3 = 1.0  # decrease query_volume_mm3 from 10 to speed up
 
 MIN_LEN_DEFECTS = 50
 ANALYZE_MAX = False
-BACKEND = "sklearn"  # "sable"
+BACKEND = "sable"
 
 MESHGRID_SIZE = 150
 
@@ -354,14 +355,18 @@ class ActiveLearningOrchestrator:
     ) -> IntersectClientCallback:
         payload = None
         if operation == "initialize_workflow":
-
             # normalize and transform the output data
             y_norm, yerr_norm = y_to_unit(
                 self.dataset_y, self.dataset_yerr, self.y_scale
             )
+            # configure the output statistics and combined dataset
+            self.labels_y = ["y", "yerr"]
+            self.statistics_y = Normal(loc="y", scale="yerr")
+            initial_dataset_y = list(zip(y_norm, yerr_norm))
 
             # prior variance of the kernel (how large is uncertainty without data)
-            prior_variance = 2.0
+            prior_std = 1.5
+            prior_variance = prior_std**2
 
             self.backend = BACKEND
             if self.backend == "sklearn":
@@ -370,21 +375,15 @@ class ActiveLearningOrchestrator:
                 length_scale = 0.2
                 self.kernel_args = {
                     "length_scale": length_scale,
-                    "length_scale_bounds": "fixed",
                     "constant_value": prior_variance,
-                    "constant_value_bounds": "fixed",
-                    # set "noise_level" (nugget) to zero, and use alpha below for heteroscedastic noise
-                    "noise_level": 0.0,
-                    "noise_level_bounds": "fixed",
                 }
-                # use the nondimensionalized yerr to set alpha
-                y_variance = np.asarray(yerr_norm) ** 2
-                self.backend_args = {"alpha": y_variance}
+                self.backend_args = {}
 
             elif self.backend == "sable":
                 self.kernel = "rbf"
                 self.kernel_args = {
                     # x range of the data
+                    # the bounds are always [0, 1], since dial currently normalizes the input
                     "x_range": self.bounds_unit[0],
                     # sigma range of valid lengthscales
                     "sigma_range": [1e-3, 0.5],
@@ -397,20 +396,20 @@ class ActiveLearningOrchestrator:
                     # memory size for number of features:
                     # needs to be large enough, but becomes slower with more features
                     "n_features": 10000,
-                    # prior variance (scaled by problem specific hyperparameter)
-                    "alpha": 0.02 / prior_variance,
+                    # prior standard deviation
+                    "prior_std": prior_std,
                     # algorithm hyperparameters
                     # p is degree of adaptivity (p=2 is a GP, p=1 is fully sparse)
                     "p": 1.25,
                     # number of optimization steps (needs to be large enough, but slows performance)
                     "n_iter_irls": 100,
-                    # noise level of the data, standard deviation of each data point
-                    "noise_level": yerr_norm,
                 }
 
[…]
