### 2026-07-17 9989ea7
fixed tests for CI, implementation of max_modes padding to account for new implementation
