### 2026-07-17 0a9456b
pre-commit formatting
