### 2026-07-17 828f065
changes to autoselection of melt pool filter
