+++ tests/test_api.py
@@ -28,6 +28,7 @@ from pathlib import Path
 from raptor.api import (
     create_grid,
     create_path_vectors,
+    compute_required_modes,
     compute_spectral_components,
     create_melt_pool,
     compute_porosity,
@@ -88,7 +89,7 @@ def sample_process_parameters():
 @pytest.fixture
 def sample_time_series_data():
     """Fixture providing sample time series data for melt pool."""
-    t = np.linspace(0, 1, 100)
+    t = np.arange(100, dtype=np.float64) / 100.0
     values = 0.0001 + 0.00002 * np.sin(2 * np.pi * 5 * t)
     return np.column_stack([t, values])
 
@@ -335,8 +336,8 @@ class TestComputeSpectralComponents:
     def test_compute_spectral_components_preserves_standard_deviation(
         self, sample_time_series_data
     ):
-        """Test that the truncated cosine representation preserves variance."""
-        spectral_array = compute_spectral_components(sample_time_series_data, 3)
+        """Test correct amplitude when the signal frequency is retained."""
+        spectral_array = compute_spectral_components(sample_time_series_data, 6)
         time = sample_time_series_data[:, 0]
         reconstructed = np.zeros_like(time)
 
@@ -348,21 +349,57 @@ class TestComputeSpectralComponents:
         )
         assert reconstructed.std() == pytest.approx(sample_time_series_data[:, 1].std())
 
-    def test_compute_spectral_components_selects_dominant_modes(self):
-        """Test that high-frequency signal content is not discarded."""
+    def test_explicit_mode_count_retains_low_frequency_prefix(self):
+        """Explicit mode counts preserve the original truncation semantics."""
         sampling_frequency = 10_000.0
         time = np.arange(1000) / sampling_frequency
         values = 1.0 + 0.2 * np.sin(2.0 * np.pi * 2000.0 * time)
         spectral_array = compute_spectral_components(np.column_stack([time, values]), 2)
 
-        assert spectral_array[1, 1] == pytest.approx(2000.0)
+        assert spectral_array[1, 1] == pytest.approx(10.0)
+
+    def test_tolerance_selects_minimum_modes_and_meets_rmse(self):
+        sampling_frequency = 1000.0
+        time = np.arange(1000) / sampling_frequency
+        values = (
+            2.0
+            + 0.30 * np.cos(2.0 * np.pi * 20.0 * time + 0.2)
+            + 0.04 * np.cos(2.0 * np.pi * 170.0 * time - 0.4)
+        )
+        data = np.column_stack([time, values])
+
+        spectral_array = compute_spectral_components(data, tolerance=0.03)
+        reconstructed = sum(
+            amplitude * np.cos(2.0 * np.pi * frequency * time + phase)
+            for amplitude, frequency, phase in spectral_array
+        )
+
+        assert spectral_array.shape == (2, 3)
+        assert spectral_array[1, 1] == pytest.approx(20.0)
+        assert np.sqrt(np.mean((values - reconstructed) ** 2)) <= 0.03
+        assert compute_required_modes(data, 0.03) == 2
+
+    def test_tolerance_zero_reconstructs_even_length_signal(self):
+        time = 0.25 + np.arange(100) / 100.0
+        values = 3.0 + 0.2 * np.cos(2.0 * np.pi * 5.0 * time)
+        values += 0.1 * np.cos(2.0 * np.pi * 50.0 * time)
+        spectral_array = compute_spectral_components(
+            np.column_stack([time, values]), tolerance=0.0
+        )
+        reconstructed = sum(
+            amplitude * np.cos(2.0 * np.pi * frequency * time + phase)
+            for amplitude, frequency, phase in spectral_array
+        )
[…]
