+++ examples/synthetic_api_example/run_synthetic_rve.py
@@ -9,11 +9,11 @@
 # https://github.com/ORNL-MDF/Raptor/LICENSE
 # =============================================================================
 import numpy as np
-import matplotlib.pyplot as plt
 
 from raptor.api import (
     compute_morphology,
     compute_porosity,
+    compute_spectral_components,
     create_grid,
     create_melt_pool,
     create_path_vectors,
@@ -21,7 +21,11 @@ from raptor.api import (
     write_morphology,
     write_vtk,
 )
-from raptor.utilities import MeltPoolFilter
+from raptor.utilities import (
+    MeltPoolFilter,
+    plot_melt_pool_signal,
+    reconstruct_spectral_signal,
+)
 
 
 # 0. User defined process parameters
@@ -37,101 +41,16 @@ MELT_POOL_HEIGHT = 30.0e-6
 MELT_POOL_WIDTH_STD_DEV = 18.0e-6
 MELT_POOL_LENGTH = 300.0e-6
 
-VARIANCE_SIGNIFICANCE = 0.025  # \alpha value for significance of variance test
-CI_TOLERANCE_WINDOW = 0.01  # relative tolerance window around the user supplied std
-
-N_SPECTRAL_MODES = 50
 HEIGHT_SHAPE_FACTOR = 1.0
 DEPTH_SHAPE_FACTOR = 1.0
 
 RVE_MIN_POINT = np.array([0.0, 0.0, 0.0])
-RVE_MAX_POINT = np.array([10.0e-4, 10.0e-4, 5.0e-4])
+RVE_MAX_POINT = np.array([5.0e-4, 5.0e-4, 5.0e-4])
 VOXEL_RESOLUTION = 5.0e-6
 
 VTK_OUTPUT = "rve.vti"
 MORPHOLOGY_OUTPUT = "rve_morphology.csv"
 WIDTH_DATA_PLOT = "melt_pool_width_timeseries.png"
-
-
-def plot_width_data(width_data):
-    time_ms = width_data[:, 0] * 1.0e3
-    width_um = width_data[:, 1] * 1.0e6
-    mean_um = MELT_POOL_WIDTH * 1.0e6
-    std_dev_um = MELT_POOL_WIDTH_STD_DEV * 1.0e6
-
-    plot_limits = (
-        min(width_um.min(), mean_um - 4.0 * std_dev_um),
-        max(width_um.max(), mean_um + 4.0 * std_dev_um),
-    )
-    gaussian_width = np.linspace(*plot_limits, 500)
-    gaussian_density = np.exp(-0.5 * ((gaussian_width - mean_um) / std_dev_um) ** 2) / (
-        std_dev_um * np.sqrt(2.0 * np.pi)
-    )
-
-    style = {
-        "font.size": 8,
-        "axes.labelsize": 8,
-        "legend.fontsize": 7,
-        "xtick.labelsize": 7,
-        "ytick.labelsize": 7,
-    }
-    with plt.rc_context(style):
-        fig, (time_axis, distribution_axis) = plt.subplots(
-            1,
-            2,
-            figsize=(6.5, 3),
-            constrained_layout=True,
-        )
-
-        time_axis.plot(time_ms, width_um, color="#0072B2", linewidth=0.8)
-        time_axis.axhline(
-            mean_um,
-            color="black",
-            linestyle="--",
-            linewidth=1.0,
-            label="Specified mean",
-        )
-        time_axis.set_xlabel("Time (ms)")
-        time_axis.set_ylabel("Melt-pool width (µm)")
-        time_axis.legend(frameon=False, loc="upper right")
-
-        distribution_axis.hist(
-            width_um,
-            bins=40,
-            density=True,
-            color="#56B4E9",
-            edgecolor="white",
[…]

[…]
