### 2026-07-20 f437106
add an output scaler that focusses attention on a range of interest
