+++ examples/active_learning_example/defect_model/create_defect_model.py
@@ -6,14 +6,13 @@ import sys
 import random
 from pathlib import Path
 from typing import Any
+from dataclasses import dataclass
 
 import numpy as np
 from scipy.stats import qmc
 from scipy.interpolate import RegularGridInterpolator
 
 # Raptor Imports
-from numba import njit
-from raptor.io import read_data
 from raptor.api import (
     create_grid,
     create_melt_pool,
@@ -24,7 +23,7 @@ from raptor.utilities import ScanPathBuilder, MeltPoolFilter
 
 # Intersect Imports
 from intersect_sdk import (
-    INTERSECT_JSON_VALUE,
+    INTERSECT_RESPONSE_VALUE,
     HierarchyConfig,
     IntersectClient,
     IntersectClientCallback,
@@ -60,17 +59,17 @@ UNIT_BOUNDS = ((0.0, 1.0),)
 NUM_DIMS = len(BOUNDS)
 
 INITIAL_DATA_SIZE = 1
-MAX_ITERATIONS = 20
+MAX_ITERATIONS = 40
 
 SEED = 42
 
 VOXEL_RESOLUTION_M = 5e-6
 RVE_LENGTH_M = 1e-3
-QUERY_VOLUME_MM3 = 1.0  # decrease query_volume_mm3 from 10 to speed up
+QUERY_VOLUME_MM3 = 3.0  # decrease query_volume_mm3 from 10 to speed up
 
 MIN_LEN_DEFECTS = 50
 ANALYZE_MAX = False
-BACKEND = "sable"
+BACKEND = "sable"  # "sable" or "sklearn"
 
 MESHGRID_SIZE = 150
 
@@ -242,7 +241,8 @@ def process_raptor_data(raptor_data):
 
     logger.info(
         f"Found {len(combined_defects)} defects: "
-        f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Max Pore: {max_pore*1e6:.2f}um"
+        f"Hatch: {raptor_data['inputs']['hatch_spacing_m']*1e6:.1f}um | Mean and Max Pore: {mean_pore*1e6:.2f}, {max_pore*1e6:.2f}um, "
+        f"Learning {'Max' if ANALYZE_MAX else 'Mean'}."
     )
 
     # TODO: discuss and refine the data analysis and extreme value statistics
@@ -284,25 +284,68 @@ def get_data_point(x_suggested, mp_interpolator):
     return x_, y_, yerr_, raptor_data
 
 
-def y_to_unit(y, yerr, y_scale=(1.0, 1.0)):
-    y_prescale, y_postscale = y_scale
-    # transformation to normalized output for training
-    #   y -> log(1 + y/y_prescale) / y_postscale
-    # scale the errors according to the derivative of the transformation
-    y = np.asarray(y, dtype=float)
-    yerr = np.asarray(yerr, dtype=float)
-    y_norm = np.log1p(y / y_prescale) / y_postscale
-    yerr_norm = yerr / (y + y_prescale) / y_postscale
-    return y_norm.tolist(), yerr_norm.tolist()
+# The unlist_list transformer
+def unlist_list(func):
+    def wrapper_func(self, *args):
+        args_np = [np.asarray(y, dtype=float) for y in args]
+        res_np = func(self, *args_np)
+        return [y.tolist() for y in res_np]
+
+    return wrapper_func
 
 
-def y_from_unit(y_norm, yerr_norm, y_scale=(1.0, 1.0)):
-    y_prescale, y_postscale = y_scale
-    y_norm = np.asarray(y_norm, dtype=float)
-    yerr_norm = np.asarray(yerr_norm, dtype=float)
[…]
